### 2022-08-10 #4665 @editorialbot
[PRE REVIEW]: NuclearToolkit.jl: A Julia package for nuclear structure calculations
**Submitting author:** @SotaYoshida (<a href="http://orcid.org/0000-0002-1342-1846">Sota Yoshida</a>)
**Repository:** https://github.com/SotaYoshida/NuclearToolkit.jl
**Branch with paper.md** (empty if default branch): josspaper
**Version:** v0.1.0
