### 2022-08-10 #4665 @danielskatz
@SotaYoshida - thanks for your submission.
Due to a current shortage of editors (that we are working to resolve), I'm going to put it on our waiting list until an appropriate editor is available.

While we wait, you could work on the invalid DOIs that editorialbot suggests. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
