### 2022-08-15 #4665 @rkurchin
@danielskatz this is slightly out of my area (I do electronic structure stuff, but don't really deal with nuclear structure), but given I know we're short on editors with Julia expertise, I could take it as I'm about to have another one accepted.
