### 2022-08-15 #4665 @mdavezac
I would be happy to, though someone with experience in nuclear structure might be a better fit (my expertise, if any, is also with electronic structure).
