### 2022-08-15 #4665 @rkurchin
No problem, George, thanks for the quick reply!

@mdavezac, our list is a bit short (read: completely lacking AFAICT) on reviewers with nuclear structure expertise who are also Julia users, so I might just invite one or two folks who know nuclear structure but don't list Julia as a language and hope we can get some complementarity with folks like yourself who use Julia and certainly at least know quantum mechanics. Hopefully that sounds okay to you, and thanks for your willingness!
