### 2022-08-15 #4665 @rkurchin
👋 @munkm and/or @villaa, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I recognize neither of you lists Julia as a language, but we're somewhat short on folks with nuclear structure expertise to review this, see discussion in prior comments...
