### 2022-08-21 SotaYoshida/NuclearToolkit.jl#23 @mdavezac
reference to global calc_3N
I'm guessing the line below should not reference some global value, but rather chiEFTobj?

https://github.com/SotaYoshida/NuclearToolkit.jl/blob/e77863fa1adf9ea884ea17a54d588763ec74f93e/src/chiEFTint/main_chiEFTint.jl#L69

Referencing openjournals/joss-reviews#4665.
