### 2022-08-21 #4665 @mdavezac
@rkurchin, if I remember the JOSS review process well, a [REVIEW] issue will be opened at some point? Or should it go here?
