### 2022-08-22 #4694 @editorialbot
[REVIEW]: NuclearToolkit.jl: A Julia package for nuclear structure calculations
**Submitting author:** @SotaYoshida (<a href="http://orcid.org/0000-0002-1342-1846">Sota Yoshida</a>)
**Repository:** https://github.com/SotaYoshida/NuclearToolkit.jl
**Branch with paper.md** (empty if default branch): main
**Version:** 0.2.5
