### 2022-08-22 #4694 @SotaYoshida
@rkurchin Thank you for your kind help and editorial work.
I have a question:
If a reviewer points out a bug or typo in the code, is it acceptable to update the code during the review process and should it be tagged?
