### 2022-08-22 SotaYoshida/NuclearToolkit.jl#23 @SotaYoshida
@mdavezac Thank you very much for your time and reviewing the code.

Absolutely you are right.
If it is OK to modify the code during the review process, I will correct it.
Related question at our [REVIEW] [here] (https://github.com/openjournals/joss-reviews/issues/4694#issuecomment-1221663351)

Thanks.
