### 2022-08-22 SotaYoshida/NuclearToolkit.jl#23 @mdavezac
My understanding is that making the code and software better is the point of the review process. So that would be an emphatic yes.
