### 2022-08-22 #4694 @rkurchin
@SotaYoshida it is not only acceptable, it is highly encouraged! As for tagging, we will ask you to issue a new release tag at the end of the review process. If you want to do so along the way, that's fine, but it's not necessary.
