### 2022-08-25 SotaYoshida/NuclearToolkit.jl#23 @SotaYoshida
@mdavezac Thanks.
I have fixed the bug (and made some improvements in terms of performance). #24
