+++ src/chiEFTint/pionexchange.jl
@@ -282,25 +282,6 @@ function tpe(chiEFTobj,LECs,ts,ws,xr,V12mom,dict_numst,to,llpSJ_s,lsjs,tllsj,opf
             tpe_for_givenJT(chiEFTobj,LoopObjects,Fpi2,tmpLECs,
                              J,pnrank,ts,ws,fff,dwn,nd_mpi,xr,pjs_para,gis_para,opfs_para,
                              f_idx,tVs_para,lsj,tllsj_para,tdict,V12mom,tmpsum,to)
-
-            # @inbounds for i= 1:n_mesh
-            # #@inbounds @threads for i= 1:n_mesh
-            #     tVs = zeros(Float64,6)
-            #     x = xr[i];xdwn = x*dwn
-            #     xdwn2 = xdwn^2       
-            #     ex = sqrt(1.0+xdwn2)
-            #     @inbounds for j = 1:n_mesh
-            #         y = xr[j];ydwn = y*dwn;ydwn2=ydwn^2
-            #         ey = sqrt(1.0+ydwn2)
-            #         k2=0.5*(xdwn2 + ydwn2)
-            #         ree = 1.0/sqrt(ex*ey)
-            #         fc = fff * hc3 * freg(x,y,2) * ree
-            #         single_tpe(chiEFTobj,LoopObjects,nd_mpi,nd_mpi2,Fpi2,
-            #                    c1,c2,c3,c4,r_d12,r_d3,r_d5,r_d145,
-            #                    J,pnrank,ts,ws,dwn,xdwn,ydwn,xdwn2,ydwn2,k2,pjs,
-            #                    gis,opfs,fc,f_idx,tVs,lsj,tllsj,tdict,V12mom,i,j,tmpsum[threadid()],to)                   
-            #     end
-            # end
         end
     end
     return nothing
@@ -750,88 +731,10 @@ end
 
 
 """
-    single_tpe(nd_mpi,nd_mpi2,Fpi2,Fpi4,Fpi6,c1,c2,c3,c4,r_d12,r_d3,r_d5,r_d145,J,pnrank,ts,ws,xdwn,ydwn,xdwn2,ydwn2,k2,pjs,gis,opfs,fc,f_idx,tVs,lsj,tllsj,tdict,V12mom,V_i,V_j,to)
+    tpe_for_givenJT(chiEFTobj,LoopObjects,Fpi2,tmpLECs,J,pnrank,ts,ws,fff,dwn,nd_mpi,xr,pjs_para,gis_para,opfs_para,f_idx,tVs_para,lsj,tllsj_para,tdict,V12mom,tmpsum,to;calc_TPE_sep=true)
 
-TPE contribution in a given momentum mesh point
+Calculating TPE contribution in a given momentum mesh point.
 """
-function single_tpe(chiEFTobj,LoopObjects,nd_mpi,nd_mpi2,Fpi2,
-                    c1,c2,c3,c4,r_d12,r_d3,r_d5,r_d145,
-                    J,pnrank,ts,ws,dwn,xdwn,ydwn,xdwn2,ydwn2,k2,pjs,
-                    gis,opfs,fc,f_idx,tVs,lsj,tllsj,tdict,V12mom,V_i,V_j,target,to;calc_TPE_sep=true)
-    LamSFR_nd = dwn * chiEFTobj.LambdaSFR
[…]
