### 2022-08-28 SotaYoshida/NuclearToolkit.jl#26 @mdavezac
Do not ignore .jl files in .gitignore
The entry `*/.jl` in `.gitignore` files makes it more difficult to navigate the code since many tools (for instance ripgrep, docker, etc...) will ignore these files. I would recommend removing the `*/.jl` entry so the code can easily be searched with standard tools.

Github recommends [this file](https://github.com/github/gitignore/blob/main/Julia.gitignore) (though for research reproducibility purposes, the last line excluding Manifest.toml is probably a bad idea).

There's also an [opt-in approach](https://www.howtogeek.com/devops/how-to-set-up-gitignore-as-a-whitelist/) that might more closely reproduce the current setup.

This is **not** a _requirement_ for openjournals/joss-reviews#4694. Please do close the issue if you disagree with the recommendation ;).
