### 2022-08-28 #4694 @mdavezac
The package is a truly ambitious effort to translate a number of nuclear structure many-body methods to Julia, with the goal of providing a modular framework to advance such methods for students and researchers. Unfortunately, the code makes little use of best practices in software design (unit-testing, separation of concerns, encapsulation, etc), Julia's capabilities promoting modularity (multiple dispatch), and reads more like a direct translation of an old Fortran code than a modern implementation. This in itself should not stop publication in JOSS. However, I believe the state of the documentation and the testing is not yet sufficient to verify the functionality claims:

1. there is no example or test for MPI (the code claims to work on supercomputers)
2. the integration tests such as they are only check the code runs to completion. They do not check the code computes anything, much less that it computes anything correctly.
3. from the input, it seems the code generates a binary output file. As far as I can tell, there is no documentation of what it contains or how to read and use the data therein.
4. the API documentation is too often sparse, especially when one considers some functions listed there take 10-20 arguments with no obvious typing. Since this package is meant as a toolkit for the community, I would recommend (though not require) investing some time into it. In this respect, unit-tests demonstrating usage and expected outputs would have been valuable.

Also, the code will not run as a package (see SotaYoshida/NuclearToolkit.jl#27).

There are a few typos in the paper:

1. first sentence, "among nucleons, nuclear force" -> "among nucleons, [the] nuclear force"
2. second sentence, "especial[l]y"
