+++ src/chiEFTint/calibration.jl
@@ -1,67 +1,30 @@
-#using PyCall
-#@pyimport matplotlib.pyplot as plt
-function myCholesky!(tmpA,ln,cLL)
-    l11 = sqrt(tmpA[1,1]) 
-    cLL[1,1] = l11
-    cLL[2,1] = tmpA[2,1]/l11; cLL[2,2] = sqrt( tmpA[2,2]-cLL[2,1]^2)
-    for i=3:ln
-        for j=1:i-1
-            cLL[i,j] = tmpA[i,j]
-            for k = 1:j-1
-                cLL[i,j] += - cLL[i,k]*cLL[j,k]                
+function caliblating_2n3nLECs_byHFMBPT(itnum,optimizer,MPIcomm,chiEFTobj,OPTobj,d9j,HOBs,nucs,HFdata,to,io;Operators=[""])    
+    @assert optimizer != "" "optimizer should be BayesOpt/LHS/MCMC"
+    if !MPIcomm        
+        for it = 1:itnum
+            ## If you want to optimize (or try samplings) change itnum, insert a function to update/optimize/sample the LECs here     
+            add2n3n(chiEFTobj,to,it)
+            @timeit to "Vtrans" dicts_tbme = TMtrans(chiEFTobj,to;writesnt=false)
+            print_vec("it = "*@sprintf("%8i",it),OPTobj.params,io)
+            @timeit to "HF/HFMBPT" hf_main_mem(chiEFTobj,nucs,dicts_tbme,d9j,HOBs,HFdata,to;Operators=Operators)
+            if optimizer=="BayesOpt"
+                BO_HFMBPT(it,OPTobj,HFdata,to)
+            elseif optimizer=="LHS"
+                LHS_HFMBPT(it,OPTobj,HFdata,to)
+            elseif optimizer=="MCMC"
+                MCMC_HFMBPT(it,OPTobj,HFdata,to)
+            else
+                @error "unknown optimizer: $optimizer " 
+                exit()
             end
-            cLL[i,j] = cLL[i,j] / cLL[j,j]
+            updateLECs_in_chiEFTobj!(chiEFTobj,OPTobj.targetLECs,OPTobj.params)
         end
-        cLL[i,i] = tmpA[i,i]
-        for j=1:i-1
-            cLL[i,i] += -cLL[i,j]^2
-        end
-        cLL[i,i] = sqrt(cLL[i,i])             
+    else
+        mpi_hfmbpt(itnum,chiEFTobj,OPTobj,d9j,HOBs,nucs,HFdata,to,io;Operators=Operators)
     end
     return nothing
 end
 
-function eval_HFMBPT(it,OPTobj,HFdata,varE,Lam;mcmc=true,io=stdout,debug=false)
-    thist = OPTobj.history[it]
-    params = OPTobj.params
-    params_ref = OPTobj.params_ref
-    tvec = params-params_ref
-    logprior = -0.5*Lam*dot(tvec,tvec)
-    if mcmc
-        Lamt = params-params_ref
-        tvec[1] *= 4.0
-        logprior = -0.5*dot(Lamt,tvec)
-    end    
-    llh = 0.0
-    for (n,tmp) in enumerate(HFdata)
-        nuc = tmp.nuc
-        data = tmp.data
-        dtype = tmp.datatype
-        A= nuc.A
-        for (i,tdtype) in enumerate(dtype)
-            vtho = data[i][1] *ifelse(tdtype=="E",1/A,1.0)
-            vexp = data[i][2] *ifelse(tdtype=="E",1/A,1.0)
-            tllh = 0.5 * (vtho-vexp)^2 / varE
-            llh -= tllh
-        end 
-    end
-    if debug; llh = 0.0;end
[…]
