+++ src/chiEFTint/pionexchange.jl
@@ -1,10 +1,14 @@
 """
     OPEP(ts,ws,xr,V12mom,dict_numst,to,lsjs,llpSJ_s,tllsj,opfs;pigamma=true)
 
-calc. One-pion exchange potential
+calc. One-pion exchange potential in the momentum-space
 """
-function OPEP(chiEFTobj,ts,ws,xr,V12mom,dict_numst,to,lsjs,llpSJ_s,tllsj,opfs,QLdict;pigamma=true)
-    n_mesh = chiEFTobj.n_mesh
+function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
+    ts = chiEFTobj.ts; ws = chiEFTobj.ws; xr = chiEFTobj.xr; V12mom = chiEFTobj.V12mom
+    dict_numst = chiEFTobj.dict_numst; 
+    lsjs = chiEFTobj.lsjs; tllsj = chiEFTobj.tllsj
+    opfs = chiEFTobj.opfs; QLdict = chiEFTobj.util_2n3n.QWs.QLdict
+    n_mesh = chiEFTobj.params.n_mesh
     hc3 = hc^3
     tVs = zeros(Float64,6)
     opfs = zeros(Float64,8)
@@ -17,7 +21,6 @@ function OPEP(chiEFTobj,ts,ws,xr,V12mom,dict_numst,to,lsjs,llpSJ_s,tllsj,opfs,QL
         coeff = -(MN*gA/(2*Fpi))^2
         itt = itts[pnrank]
         tllsj[1] = itt
-        TF = true
         @inbounds for J=0:jmax
             lsj = lsjs[J+1]
             f_idx = 6
@@ -53,6 +56,18 @@ function OPEP(chiEFTobj,ts,ws,xr,V12mom,dict_numst,to,lsjs,llpSJ_s,tllsj,opfs,QL
                     end
                 end
             end
+            ### check
+            if debugmode 
+                @inbounds for idx = 1:f_idx
+                    @views tllsj[2:5] .= lsj[idx]
+                    tl,tlp,tS,tJ = lsj[idx] 
+                    if pnrank%2 == 1 && (tl+tS+1)%2 != 1;continue;end
+                    V12idx = get(tdict,tllsj,-1)
+                    if V12idx == -1;continue;end                        
+                    V = V12mom[V12idx]
+                    #println("pnrank $pnrank J $J V12idx $V12idx");println(V)
+                end
+            end 
         end
     end
     return nothing
@@ -143,7 +158,7 @@ function cib_lsj_opep(opfs,x,y,mpi2,nterm,J,pnrank,facin,ts,ws,tVs,QLdict,pigamm
     return nothing 
 end
 
-### function-forms for pw
+### function-forms for partial waves
 f1(x,y,c,cdum) = c
 fx(x,y,c,cdum) = c * x
 fy(x,y,c,cdum) = c * y
@@ -159,7 +174,6 @@ f_x3y(x,y,c,cdum) = c * x^3 * y
 f_x3y3(x,y,c,cdum) = c * x^3 * y^3
 f_xy3(x,y,c,cdum) = c * x * y^3
 f_31(x,y,c,cdum) = c * (x^3 * y + x * y^3)
-
 fp_P2(p,ell,pp,ellp,P) = P^2
 fp_ddP(p,ell,pp,ellp,P) = P * (delta(ell,1)*delta(ellp,0)*p + delta(ell,0)*delta(ellp,1)*pp)
 
[…]

[…]
