+++ src/package_install.jl
@@ -1,19 +1,0 @@
-using Pkg
-
-
-targets = ["Arpack","Combinatorics","Glob",
-      "LaTeXStrings",
-      "LatinHypercubeSampling",
-      "LinearAlgebra",
-      "Printf",
-      "PyCall",
-      "Random",
-      "StatsBase",
-      "Statistics",
-      "SpecialFunctions",
-      "TimerOutputs",
-      "ThreadPools",
-      "WignerSymbols"]
-for tmp in targets
-    Pkg.add(tmp)
-end
