+++ src/chiEFTint/contact.jl
@@ -1,4 +1,5 @@
 """
+    LO(chiEFTobj,to)    
 
 Calculate Leading Order (LO), `Q^0`` contact term.
 """
@@ -6,7 +7,7 @@ function LO(chiEFTobj,to)
     xr = chiEFTobj.xr
     LECs = chiEFTobj.LECs.dLECs
     V12mom = chiEFTobj.V12mom
-    dict_numst = chiEFTobj.dict_numst
+    dict_pwch = chiEFTobj.dict_pwch
     C_1S0 = LECs["C0_1S0"]
     C_3S1 = LECs["C0_3S1"]
     C_CIB = LECs["C_CIB"]
@@ -20,12 +21,12 @@ function LO(chiEFTobj,to)
         ##1S0 (LO)
         l=0;lp=0;S=0;J=0
         LEC = (C_1S0+res) * fac_q
-        calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_numst[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
+        calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_pwch[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
         ## 3S1 (LO)
         l=0;lp=0;S=1;J=1
         if pnrank%2==1;continue;end
         LEC = (C_3S1) * fac_q 
-        calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_numst[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
+        calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_pwch[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
     end
     return nothing
 end
@@ -44,7 +45,7 @@ function NLO(chiEFTobj,to;n_regulator=2)
     xr = chiEFTobj.xr
     LECs = chiEFTobj.LECs.dLECs
     V12mom = chiEFTobj.V12mom
-    dict_numst = chiEFTobj.dict_numst
+    dict_pwch = chiEFTobj.dict_pwch
     fac = hc^3  *  1.e-8  / (2*pi)^3  
     tLECs = [ LECs["C2_3S1"],LECs["C2_3P0"],LECs["C2_1P1"],
               LECs["C2_3P1"],LECs["C2_1S0"],
@@ -63,7 +64,7 @@ function NLO(chiEFTobj,to;n_regulator=2)
             else
                 n_reg = n_regulator
             end
-            calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_numst[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
+            calc_Vmom!(chiEFTobj.params,pnrank,V12mom,dict_pwch[pnrank],xr,LEC,LEC,l,lp,S,J,pfunc,n_reg,to)
         end
     end
     return nothing
@@ -78,7 +79,7 @@ function N3LO(chiEFTobj,to)
     xr = chiEFTobj.xr
     LECs = chiEFTobj.LECs.dLECs
     V12mom = chiEFTobj.V12mom
-    dict_numst = chiEFTobj.dict_numst
+    dict_pwch = chiEFTobj.dict_pwch
     fac = hc^3  *  1.e-14  / (2*pi)^3
     LECs_N3LO = [[LECs["hD_1S0"],LECs["D_1S0"]],[LECs["D_3P0"]],
                  [LECs["D_1P1"]],[LECs["D_3P1"]],
@@ -110,17 +111,22 @@ function N3LO(chiEFTobj,to)
[…]

[…]
