+++ docs/src/howtouse.md
@@ -4,7 +4,7 @@ In this page, we explain the main interfaces of NuclearToolkit.jl and sample cod
 
 ## Generate Chiral EFT potentials
 
-The [ChiEFTint](References/ChiEFTint) subblock provides the codes to generate nucleon-nucleon (NN) potentials from Chiral effective field theory. You can specify the options through `optional_parameters.jl`, if it exists in the working directory.
+The [ChiEFTint](../ChiEFTint) subblock provides the codes to generate nucleon-nucleon (NN) potentials from Chiral effective field theory. You can specify the options through `optional_parameters.jl`, if it exists in the working directory.
 A sample of `optional_parameters.jl` is available on the [repository](https://github.com/SotaYoshida/NuclearToolkit.jl), and see the [Optional parameters](parameters) page for more details.
 
 The main API is `make_chiEFTint()` and the sample code is given below.
@@ -40,7 +40,7 @@ When working on a working node (w/o permissions to access `~/`), overwrite the `
 
 ## HF-MBPT calculations 
 
-You can evaluate ground state properties by the so-called Hartree-Fock many-body perturbation theory (HF-MBPT). The `hf_main` function in [HartreeFock](HartreeFock) is the main interface.
+You can evaluate ground state properties by the so-called Hartree-Fock many-body perturbation theory (HF-MBPT). The `hf_main` function in [HartreeFock](../HartreeFock) is the main interface.
 
 You must specify target nuclei, snt file (generated by chiEFTint), harmonic oscillator parameter `hw`, model space size `emax`.
 
