### 2022-09-10 6ed6d1f
dealing with smaller emax calc with a given interacton
