+++ test/optional_parameters_snt.jl
@@ -1,12 +1,12 @@
 ###--- ChiralEFT ---
 n_mesh = 50
 emax = 2
-chi_order = 3
+chi_order = 3 
 calc_NN = true
-calc_3N = false 
+calc_3N = false
 hw = 20.0
 srg = false
-tbme_fmt = "snt.bin"
+tbme_fmt = "snt"
 ### --- IMSRG ---
 smax = 500.0
 dsmax = 0.5
