+++ src/hartreefock.jl/main.jl
@@ -5,7 +5,7 @@ main function to carry out HF/HFMBPT calculation from snt file
 - `nucs::Vector{String}` target nuclei
 - `sntf` path to input interaction file
 - `hw` hbar omega
-- `emax` emax for HF/IMSRG
+- `emax_calc` emax for HF/IMSRG
 
 # Optional Arguments
 - `verbose=false` to see detailed stdout for HF
@@ -16,24 +16,25 @@ main function to carry out HF/HFMBPT calculation from snt file
 - `corenuc=""` core nucleus, example=> "He4"
 - `ref="nucl"` to specify target reference state, "core" or "nucl" is supported
 """
-function hf_main(nucs,sntf,hw,emax;verbose=false,Operators=String[],is_show=false,doIMSRG=false,valencespace=[],corenuc="",ref="nucl",io=stdout)
+function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show=false,doIMSRG=false,valencespace=[],corenuc="",ref="nucl",io=stdout,return_HFobj=false)
     @assert isfile(sntf) "sntf:$sntf is not found!"
     to = TimerOutput()
     chiEFTparams = init_chiEFTparams(;io=io)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc)
     @timeit to "PreCalc 6j" begin
-        dict6j,d6j_nabla,d6j_int = PreCalc6j(emax)        
+        dict6j,d6j_nabla,d6j_int = PreCalc6j(emax_calc)   
     end
-    @timeit to "PreCalc 9j&HOBs" d9j,HOBs = PreCalcHOB(chiEFTparams,d6j_int,to)
+    @timeit to "PreCalc 9j&HOBs" d9j,HOBs = PreCalcHOB(chiEFTparams,d6j_int,to;emax_calc=emax_calc)
     @timeit to "read" begin        
         TF = occursin(".bin",sntf)
         tfunc = ifelse(TF,readsnt_bin,readsnt)     
         nuc = def_nuc(nucs[1],ref,corenuc)
-        binfo = basedat(nuc,sntf,hw,emax,ref)
+        binfo = basedat(nuc,sntf,hw,emax_calc,ref)
         sps,dicts1b,dicts = tfunc(sntf,binfo,to)
         Z = nuc.Z; N=nuc.N; A=nuc.A
-        BetaCM = chiEFTparams.BetaCM       
+        BetaCM = chiEFTparams.BetaCM
         Hamil,dictsnt,Chan1b,Chan2bD,Gamma,maxnpq = store_1b2b(sps,dicts1b,dicts,binfo)
+
         HCM = InitOp(Chan1b,Chan2bD.Chan2b)
         TCM = InitOp(Chan1b,Chan2bD.Chan2b)
         VCM = InitOp(Chan1b,Chan2bD.Chan2b)
@@ -44,7 +45,6 @@ function hf_main(nucs,sntf,hw,emax;verbose=false,Operators=String[],is_show=fals
             aOp!(VCM,fac_HCM)            
             aOp1_p_bOp2!(VCM,HCM,1.0,0.0)
             CalculateTCM!(TCM,binfo,Chan1b,Chan2bD.Chan2b,sps)
-            #aOp1_p_bOp2!(TCM,HCM,1.0/A,1.0)
             update_dicts_withHCM!(HCM,Chan2bD,dicts)
         end 
         MatOp = Matrix{Float64}[]
@@ -56,25 +56,22 @@ function hf_main(nucs,sntf,hw,emax;verbose=false,Operators=String[],is_show=fals
     for (i,tnuc) in enumerate(nucs)
[…]
