+++ src/chiEFTint/main_chiEFTint.jl
@@ -14,7 +14,7 @@ The function is exported and can be simply called make_chiEFTint() in your scrip
 - `MPIcomm::Bool`, to carry out LECs sampling with HF-MBPT and affine inveriant MCMC
 - `Operators::Vector{String}` specifies operators you need to use in LECs calibrations
 """
-function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[])
+function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[],fn_params="optional_parameters.jl")
     to = TimerOutput()
     io=stdout
     optHFMBPT=false
@@ -32,7 +32,7 @@ function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="
         io = open("logfile.dat","w")
     end
     
-    @timeit to "prep." chiEFTobj,OPTobj,d9j,HOBs = construct_chiEFTobj(optHFMBPT,itnum,optimizer,MPIcomm,io,to)
+    @timeit to "prep." chiEFTobj,OPTobj,d9j,HOBs = construct_chiEFTobj(optHFMBPT,itnum,optimizer,MPIcomm,io,to;fn_params)
     @timeit to "NNcalc" calcualte_NNpot_in_momentumspace(chiEFTobj,to)
     @timeit to "renorm." SRG(chiEFTobj,to)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc) 
@@ -85,8 +85,8 @@ j_a/2&  j_b/2&     1 \\\\
 - `ts::Vector{Float64}` mesh points for angular integrals in pion-exchange terms
 - `ws::Vector{Float64}` weights vectors for angular integrals in pion-exchange terms
 - `Numpn::Dict{Vector{Int64},Int64}`
-- `infos::Vector{Vector{Int64}}`
-- `izs_ab::Vector{Vector{Vector{Int64}}}`
+- `infos::Vector{Vector{Int64}}` information of two-body channeles: { [``T_z``,parity,J,dim] } 
+- `izs_ab::Vector{Vector{Vector{Int64}}}` two-body kets: { [iza,ia,izb,ib] } where `iz*` stuff is isospin (-1 or 1) and `i*` stuff is label of sps. For example, [-1,1,1,1] corresponds to |p0s1/2 n0s1/2>.
 - `nTBME::Int64` # of two-body matrix elements (TBMEs)
 - `util_2n3n::util_2n3n`
 - `LECs::LECs`
@@ -139,9 +139,9 @@ It returns
 - `d9j::Vector{Vector{Vector{Vector{Vector{Vector{Vector{Float64}}}}}}}` array of Wigner-9j symbols used for Pandya transformation in HF-MBPT/IMSRG calculations.
 - `HOBs::Dict{Int64, Dict{Int64, Float64}}` dictionary for harmonic oscillator brackets.
 """
-function construct_chiEFTobj(optHFMBPT,itnum,optimizer,MPIcomm,io,to)
+function construct_chiEFTobj(optHFMBPT,itnum,optimizer,MPIcomm,io,to;fn_params="optional_parameters.jl")
     # specify chiEFT parameters
-    params = init_chiEFTparams(;io=io)
+    params = init_chiEFTparams(;io=io,fn_params=fn_params)
 
[…]
