### 2022-09-12 6609bb7
performance improved for calclations with truncated emax
