+++ src/hartreefock.jl/main.jl
@@ -16,11 +16,11 @@ main function to carry out HF/HFMBPT calculation from snt file
 - `corenuc=""` core nucleus, example=> "He4"
 - `ref="nucl"` to specify target reference state, "core" or "nucl" is supported
 """
-function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show=false,doIMSRG=false,valencespace=[],corenuc="",ref="nucl",return_HFobj=false)
+function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show=false,doIMSRG=false,valencespace=[],corenuc="",ref="nucl",return_HFobj=false,oupfn="")
     @assert isfile(sntf) "sntf:$sntf is not found!"
     to = TimerOutput()
-    io = open("logfile.dat","w")
-    chiEFTparams = init_chiEFTparams(;io=io)
+    io = select_io(false,"",nucs;use_stdout=true,fn=oupfn)
+    chiEFTparams = init_chiEFTparams(;io=nothing)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc)
     @timeit to "PreCalc 6j" begin
         dict6j,d6j_nabla,d6j_int = PreCalc6j(emax_calc)   
@@ -31,7 +31,7 @@ function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show
         tfunc = ifelse(TF,readsnt_bin,readsnt)     
         nuc = def_nuc(nucs[1],ref,corenuc)
         binfo = basedat(nuc,sntf,hw,emax_calc,ref)
-        @time sps,dicts1b,dicts = tfunc(sntf,binfo,to)
+        sps,dicts1b,dicts = tfunc(sntf,binfo,to)
         Z = nuc.Z; N=nuc.N; A=nuc.A
         BetaCM = chiEFTparams.BetaCM
         Hamil,dictsnt,Chan1b,Chan2bD,Gamma,maxnpq = store_1b2b(sps,dicts1b,dicts,binfo)
@@ -88,7 +88,6 @@ function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show
         if return_HFobj; return HFobj;end
     end
     if is_show; show(io,to, allocations = true,compact = false);println(""); end
-    close(io)
     return true
 end
 
