### 2022-09-15 1d21143
remove redundant function for HOB
