### 2022-09-16 1860f2f
redundant function&lines are removed
