+++ src/chiEFTint/pionexchange.jl
@@ -51,8 +51,7 @@ function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
                         V12idx = get(tdict,tllsj,-1)
                         if V12idx == -1;continue;end                        
                         tfac = tVs[idx] * t_fc
-                        V = V12mom[V12idx]                            
-                        V[i,j] += tfac
+                        V12mom[V12idx][i,j] += tfac
                     end
                 end
             end
@@ -65,13 +64,15 @@ function fac_pig(beta,c5=0.0)
     return - (1.0-beta)^2 / (2*beta^2) * log(1+beta) +(1.0+beta)/(2*beta) -2.0*c5
 end
 
+"""
+
+Reference: R. Machleidt, Phys. Rev. C 63 024001 (2001).
+"""
 function cib_lsj_opep(opfs,x,y,mpi2,nterm,J,pnrank,facin,ts,ws,tVs,QLdict,pigamma=false;additive=false)
     x2 = x^2; y2 = y^2
     z = (mpi2+x2+y2) / (2.0*x*y)
-    QJ   = 0.0
-    QJm1 = 0.0
+    QJ = QJm1 = 0.0
     nfac = facin
-    xdwn = x /Ms[pnrank]
     if pigamma
         nfac = facin * fsalpha/pi
         q2s = zeros(Float64,length(ts))
@@ -82,18 +83,17 @@ function cib_lsj_opep(opfs,x,y,mpi2,nterm,J,pnrank,facin,ts,ws,tVs,QLdict,pigamm
         end
         QJ = QL(z,J,q2s,ws,QLdict) 
         if J>0;QJm1=QL(z,J-1,q2s,ws,QLdict);end
-        #println("correction QJ $QJ QJm1 $QJm1")
     else
         QJ = QL(z,J,ts,ws,QLdict)
         if J>0;QJm1=QL(z,J-1,ts,ws,QLdict);end
-        #println("normal QJ $QJ QJm1 $QJm1")
     end
     IJ0 = nfac * QJ
-    IJ1 = nfac * (z * QJ -delta(J,0))
-    IJ2 = nfac * (J*z* QJ + QJm1) /(J+1) 
-    IJ3 = nfac * sqrt(J/(J+1)) * (z* QJ - QJm1)    
-    v1  = opfs[1] * IJ0 + opfs[2] *IJ1
-    v2  = opfs[3] * IJ0 + opfs[4] *IJ2
+    IJ1 = nfac * (z * QJ -delta(J,0)) #Eq. (B19)
+    IJ2 = nfac * (J*z* QJ + QJm1) /(J+1) #Eq. (B20)
+    IJ3 = nfac * sqrt(J/(J+1)) * (z* QJ - QJm1) #Eq. (B21)
+    #Eq. (B28)
+    v1 = opfs[1] * IJ0 + opfs[2] *IJ1 
+    v2 = opfs[3] * IJ0 + opfs[4] *IJ2
     v3 = opfs[5] * IJ0 + opfs[6] *IJ1
     v4 = opfs[4] * IJ0 + opfs[3] *IJ2
     v5 = opfs[7] * IJ3
@@ -106,42 +106,24 @@ function cib_lsj_opep(opfs,x,y,mpi2,nterm,J,pnrank,facin,ts,ws,tVs,QLdict,pigamm
[…]
