+++ test/optional_parameters.jl
@@ -1,12 +1,12 @@
 ###--- ChiralEFT ---
 n_mesh = 50
 emax = 2
-chi_order = 3
+chi_order = 4
 calc_NN = true
-calc_3N = false 
+calc_3N = true
 hw = 20.0
-srg = false
 tbme_fmt = "snt.bin"
+pottype="emn500n4lo"
 ### --- IMSRG ---
 smax = 500.0
 dsmax = 0.5
