+++ src/chiEFTint/eff3nf.jl
@@ -20,29 +20,10 @@ struct util_2n3n
     wsyms::wsyms_j1_1or2
 end
 
-"""
-    prep_integrals_for2n3n(chiEFTobj)
-
-preparing integrals and Wigner symbols for density-dependent 3NFs:
-- `Fis::Fis_2n3n` vectors {F0s,F1s,F2s,F3s}, Eqs.(A13)-(A16) in [Kohno2013].
-- `QWs::` second kind Legendre functions, Eqs.(B1)-(B5) in [Kohno2013].
-- `wsyms::` Wingner symbols with specific `j` used for both 2n and 2n3n.
-
-Reference:  
-[Kohno2013] M.Kohno Phys. Rev. C 88, 064005(2013)
-"""
-function prep_integrals_for2n3n(chiEFTobj,xr,ts,ws)
-    n_mesh = ifelse(calc_3N,chiEFTobj.n_mesh,3)
-    F0s = zeros(Float64,n_mesh); F1s = zeros(Float64,n_mesh); F2s = zeros(Float64,n_mesh); F3s = zeros(Float64,n_mesh)
-    QWs = prep_QWs(chiEFTobj,xr,ts,ws)
-    prep_Fis!(chiEFTobj,xr,F0s,F1s,F2s,F3s)        
-    wsyms = prep_wsyms()
-    Fis = Fis_2n3n(F0s,F1s,F2s,F3s)
-    return util_2n3n(Fis,QWs,wsyms)
+function XF(Anum)
+    return (1.0 + (0.35*pi/1.15)^2 * Anum^(-2/3) )^(-1/3)
 end
-
 function prep_Fis!(chiEFTobj,xr,F0s,F1s,F2s,F3s)
-    if !chiEFTobj.calc_3N; return nothing;end
     kF = chiEFTobj.kF
     n_mesh = chiEFTobj.n_mesh
     mpi = sum(mpis)/3.0 / hc; mpi2 = mpi^2; mpi4=mpi2^2
@@ -68,7 +49,31 @@ function prep_Fis!(chiEFTobj,xr,F0s,F1s,F2s,F3s)
     return nothing
 end
 
-function calc_vmom_3nf(chiEFTobj,it,to;pnm=false,nreg = 3)
+"""
+    prep_integrals_for2n3n(chiEFTobj)
+
+preparing integrals and Wigner symbols for density-dependent 3NFs:
+- `Fis::Fis_2n3n` vectors {F0s,F1s,F2s,F3s}, Eqs.(A13)-(A16) in [Kohno2013].
+- `QWs::` second kind Legendre functions, Eqs.(B1)-(B5) in [Kohno2013].
+- `wsyms::` Wingner symbols with specific `j` used for both 2n and 2n3n.
+
+Reference:  
+[Kohno2013] M.Kohno Phys. Rev. C 88, 064005(2013)
+"""
+function prep_integrals_for2n3n(chiEFTobj,xr,ts,ws)
+    calc_3N = chiEFTobj.calc_3N
+    n_mesh = ifelse(calc_3N,chiEFTobj.n_mesh,3)
+    F0s = zeros(Float64,n_mesh); F1s = zeros(Float64,n_mesh); F2s = zeros(Float64,n_mesh); F3s = zeros(Float64,n_mesh)
+    QWs = prep_QWs(chiEFTobj,xr,ts,ws)
+    if calc_3N
+        prep_Fis!(chiEFTobj,xr,F0s,F1s,F2s,F3s)        
+    end
+    wsyms = prep_wsyms()
+    Fis = Fis_2n3n(F0s,F1s,F2s,F3s)
+    return util_2n3n(Fis,QWs,wsyms)
+end
+
+function calc_vmom_3nf(chiEFTobj,it,to;pnm=false)
[…]
