+++ src/hartreefock.jl/main.jl
@@ -77,8 +77,10 @@ function hf_main(nucs,sntf,hw,emax_calc;verbose=false,Operators=String[],is_show
            imsrg_main(binfo,Chan1b,Chan2bD,HFobj,dictsnt,d9j,HOBs,dict6j,valencespace,Operators,MatOp,to)
         else
             if "Rp2" in Operators
-                Op_Rp2 = InitOp(Chan1b,Chan2bD.Chan2b)
-                @timeit to "Rp2" eval_rch_hfmbpt(binfo,Chan1b,Chan2bD,HFobj,Op_Rp2,d9j,HOBs,dict6j,MatOp,to;io=io)
+                @timeit to "Rp2" begin 
+                    Op_Rp2 = InitOp(Chan1b,Chan2bD.Chan2b)
+                    eval_rch_hfmbpt(binfo,Chan1b,Chan2bD,HFobj,Op_Rp2,d9j,HOBs,dict6j,MatOp,to;io=io)
+                end
             end
         end
         Aold = A
