+++ .github/workflows/CI.yml
@@ -22,7 +22,7 @@ jobs:
           - {version: '1.7',   os: ubuntu-latest,  payload: serial}
           - {version: '1.7',   os: macOS-latest,   payload: serial}
           - {version: '1,7',   os: ubuntu-latest,  payload: mpi}
-          - {version: nigthly,   os: ubuntu-latest,  payload: serial}
+          - {version: nightly,   os: ubuntu-latest,  payload: serial}
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
