+++ test/optional_parameters_snt.jl
@@ -1,13 +1,13 @@
 ###--- ChiralEFT ---
-n_mesh = 50
+n_mesh = 10
 emax = 2
-chi_order = 4
+chi_order = 3
 calc_NN = true
 calc_3N = true
 hw = 20.0
 srg = false
 tbme_fmt = "snt"
-pottype="emn500n4lo"
+pottype="em500n3lo"
 ### --- IMSRG ---
 smax = 500.0
 dsmax = 0.5
