+++ .github/workflows/CI.yml
@@ -19,7 +19,7 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - {version: '1.7',   os: ubuntu-latest,  payload: mpi}
+          - {version: 'latest',   os: ubuntu-latest,  payload: mpi}
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
