+++ .github/workflows/CI.yml
@@ -19,7 +19,7 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          - {version: 'latest',   os: ubuntu-latest,  payload: mpi}
+          - {version: 'nightly',   os: ubuntu-latest,  payload: mpi}
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
