+++ .github/workflows/CI.yml
@@ -32,8 +32,20 @@ jobs:
         uses: julia-actions/julia-runtest@v1
         if: ${{ matrix.payload == 'serial' }}
         continue-on-error: ${{ matrix.version == 'nightly' }}
-      - name: Execute MPI-parallel tests
-        run: mpirun -np 4 julia mpirun_test.jl
-        if: ${{ matrix.payload == 'mpi' }}
-        continue-on-error: ${{ matrix.version == 'nightly' }}
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        mpi: [ 'mpich', 'openmpi' ]
+    name: ${{ matrix.mpi }} example
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Setup MPI
+        uses: mpi4py/setup-mpi@v1
+        with:
+          mpi: ${{ matrix.mpi }}
+      - run: mpiexec -n 4 julia ./mpirun_test.jl
+      if: ${{ matrix.payload == 'mpi' }}
+      continue-on-error: ${{ matrix.version == 'nightly' }}
 
