+++ .github/workflows/CI.yml
@@ -37,7 +37,6 @@ jobs:
     strategy:
       matrix:
         mpi: [ 'mpich', 'openmpi' ]
-    name: MPI example
     steps:
       - name: Checkout
         uses: actions/checkout@v3
