+++ .github/workflows/CI.yml
@@ -35,6 +35,7 @@ jobs:
   build:
     runs-on: ubuntu-latest
     strategy:
+      fail-fast: false
       matrix:
         mpi: [ 'mpich', 'openmpi' ]
     steps:
