+++ .github/workflows/CI.yml
@@ -40,8 +40,13 @@ jobs:
         mpi: [ 'openmpi' ]
     name: mpi test
     steps:
-      - name: Checkout
-        uses: actions/checkout@v3
+      - uses: actions/checkout@v2
+      - uses: julia-actions/setup-julia@v1
+        with:
+          version: ${{ matrix.version }}
+          arch: x64
+      - uses: julia-actions/cache@v1
+      - uses: julia-actions/julia-buildpkg@v1
       - name: Setup MPI
         uses: mpi4py/setup-mpi@v1
         with:
