### 2022-09-18 9201da3
add deuteron func
