+++ src/chiEFTint/main_chiEFTint.jl
@@ -23,6 +23,7 @@ function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="
     io = select_io(MPIcomm,optimizer,nucs)
     @timeit to "prep." chiEFTobj,OPTobj,d9j,HOBs = construct_chiEFTobj(do2n3ncalib,itnum,optimizer,MPIcomm,io,to;fn_params)
     @timeit to "NNcalc" calcualte_NNpot_in_momentumspace(chiEFTobj,to)
+    @timeit to "deutron" Calc_Deuteron(chiEFTobj,to)
     @timeit to "renorm." SRG(chiEFTobj,to)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc) 
     if do2n3ncalib #calibrate 2n3n LECs by HFMBPT
@@ -435,8 +436,7 @@ function read_LECs(pottype)
     elseif pottype == "emn500n4lo"
         return dict_emn500n4lo()
     else
-        @error "unknown potype $pottype"
-        exit()
+        @error "unknown potype $pottype"       
     end
 end
 
@@ -475,8 +475,7 @@ function Rnl_all_ab(chiEFTobj,lmax_in,br,n_mesh,xr_fm)
     Rnl = zeros(Float64,Nnmax+1,lmax_in+1,n_mesh)
     for l=0:lmax_in
         for kidx=1:n_mesh
-            pb = br * xr_fm[kidx]
-            pb2 = pb^2
+            pb = br * xr_fm[kidx]; pb2 = pb^2
             fexp = exp(-0.5*pb2)
             fpow = pb^(l+1)
             for n=0:Nnmax
@@ -492,65 +491,34 @@ end
 """
     prepare_2b_pw_states(;io=stdout)
 
-preparing two-body channels in terms of <Lp,S,J| |L,S,J>.
+preparing two-body partical-wave channels, <Lp,S,J| |L,S,J>.
 For example, [pnrank,L,Lp,S,J] = [0, 0, 2, 1, 1] corresponds to proton-neutron 3S1-3D1 channel.
 """
 function prepare_2b_pw_states(;io=stdout)
-    #iz,lz1,lz2,isz,jz
     pw_channels = Vector{Int64}[ ]
     dict_pwch = [Dict{Vector{Int64},Int64}() for pnrank=1:3] 
     arr_pwch = [[[[ zeros(Int64,j+iss-abs(j-iss)+1) for ll1=abs(j-iss):j+iss ] for j=0:jmax ] for iss=0:1 ] for pnrank=1:3]    
     num=0
-    ## pp iz = -2
-    itt = 1
-    for iss=0:1
-        for j=0:jmax
-            for ll1=abs(j-iss):j+iss
-                for ll2=abs(j-iss):j+iss
-                    if (-1)^ll1 != (-1)^ll2;continue;end
-                    if itt != Int( (1+(-1)^(ll1+iss))/2);continue;end
-                    if (-1)^(ll1+iss+itt) != -1;continue;end
-                    num=num+1
-                    push!(pw_channels,[-2,ll1,ll2,iss,j])
-                    dict_pwch[1][[-2,ll1,ll2,iss,j]] = num
-                    arr_pwch[1][iss+1][j+1][ll1-abs(j-iss)+1][ll2-abs(j-iss)+1] = num
-                end
-            end
-        end
-    end
-    ##  nn iz = 2
-    pnrank=3
-    for iss=0:1
-        for j=0:jmax
-            for ll1=abs(j-iss):j+iss
-                for ll2=abs(j-iss):j+iss
-                    if (-1)^ll1 != (-1)^ll2;continue;end
-                    if itt != Int( (1+(-1)^(ll1+iss))/2);continue;end
[…]
