+++ src/chiEFTint/misc_plt_io.jl
@@ -7,8 +7,8 @@ hat(a) = sqrt(2.0*a+1.0)
 const jmax = 6
 const lmax = jmax +1
 const lcmax = jmax + 1
-const Mp = 938.272
-const Mn = 939.5653
+const Mp = 938.272088 
+const Mn = 939.565420 
 const Mm = (Mp+Mn)/2
 const Ms = [Mp,Mm,Mn]
 const Lambchi = 500 # cutoff Lambda
