### 2022-09-20 6b3ee3d
fixed potential bug around HOBs using multi-threading
