+++ src/IMSRG.jl/imsrg_util.jl
@@ -602,12 +602,12 @@ function need_check(tbc_bra_cc, tbc_ket_cc, li,ji,tzi, lj,jj,tzj, lk,jk,tzk, ll,
 end
 
 
-function get_nkey_from_key6j(tkey,ofst_unit=1000)
+function get_nkey_from_key6j(tkey;ofst_unit=1000)
     nkey = tkey[1] + tkey[2] * ofst_unit + tkey[3] * ofst_unit^2 + tkey[4] * ofst_unit^3 +  tkey[5] * ofst_unit^4
     return nkey
 end
-function get_nkey_from_key6j(ji,jj,jk,jl,Jp,ofst=1000)
-    return ji + jj*ofst + jk * ofst^2+ jl* ofst^3 + Jp * ofst^4
+function get_nkey_from_key6j(ji,jj,jk,jl,Jp;ofst_unit=1000)
+    return ji + jj*ofst_unit + jk * ofst_unit^2+ jl* ofst_unit^3 + Jp * ofst_unit^4
 end
 
 function IMSRGflow(binfo,HFobj,IMSRGobj,PandyaObj,Chan1b,Chan2bD,dictMono,dict6j,

[…]
