+++ test/optional_parameters.jl
@@ -1,13 +1,13 @@
 ###--- ChiralEFT ---
-n_mesh = 10
+n_mesh = 50
 emax = 2
-chi_order = 3
 calc_NN = true
-calc_3N = true
+calc_3N = false
 hw = 20.0
 srg = false
 tbme_fmt = "snt.bin"
-pottype="em500n3lo"
+chi_order = 3; pottype="em500n3lo"
+
 ### --- IMSRG ---
 smax = 500.0
 dsmax = 0.5
