+++ docs/src/parameters.md
@@ -1,6 +1,11 @@
 # Parameters in NuclearToolkit
 
-Here is the summary of optional parameters, which can be specified by users through `optional_parameters.jl` in the working directory.
+Here is the summary of optional parameters, which can be specified by users like this:
+```julia
+make_chiEFTint(;fn_params="your_parameterfile.jl")
+```
+
+If you do not specify the optional argument `fn_params`, the pacakege tries to read "optional_parameters.jl" in the current path.
 
 * For ChiEFTint
     - `n_mesh::Int64`, number of momentum mesh 
