+++ docs/src/index.md
@@ -34,14 +34,14 @@ NuclearToolkit.jl provides a self-contained set of nuclear structure calculation
 
 One of the main motivations for the author to develop NuclearToolkit.jl is, of course, for their own research purposes, and another one is for educational purpose.
 A single nuclear structure calculation code often contains tens of thousands to nearly 100,000 lines.
-In the author's personal opinion, it would be too difficult for students (especially in undergraduate or two-year master course) to understand the technical details of the latest nuclear many-body methods while reading the enormous length of existing codes in the community.
+In the author's personal opinion, it would be too difficult for students (especially in undergraduate or two-year master course) to understand the technical details of the latest nuclear many-body methods while reading the existing code with enormous length.
 
-The author thought the Julia language can be a game changer to this situation with its high readbility, portabillity, and performance. Since all the source code in NuclearToolkit.jl is a pure Julia implementation, there is no more need to prepare different Makefiles for different environments, worry about library dependencies, homemade Python script to run the Fortran/C++ codes. The code can be easily executed on either a laptop or a supercomputer.
+The author thought the Julia language can be a game changer to this situation with its high readbility, portabillity, and performance. Since all the source code in NuclearToolkit.jl is a pure Julia implementation, there is no more need to prepare different Makefiles for different environments, worry about library dependencies, homemade Python scripts to run the Fortran/C++ codes. The code can be easily executed on either a laptop or a supercomputer.
 While NuclearToolkit covers a wide range of methods, the overall code length is still in a few tens of thousands, including "docstring" to generate the document.
 
-- [ChiEFTint](ChiEFTint): NN interaction from Chiral EFT ~ 6,000 lines.
+- [ChiEFTint](ChiEFTint): NN interaction from Chiral EFT ~ 6,500 lines.
   - EM (Entem & Machleidt) N3L0
-  - EMN (Entem, Machleidt, Nosyc) N4LO
+  - EMN (Entem, Machleidt, Nosyk) N4LO
   - Density-Dependent NN from 3NF (The author prefers to call it "2n3n" to distinguish with genuine 3NF)
   - valence chiral EFT potential upto LO
 - [Hartreefock](HartreeFock): Hartree-Fock (HF) and HF Many-Body Perturbation Theory (HFMBPT)  ~ 3,000 lines.
@@ -59,10 +59,11 @@ While NuclearToolkit covers a wide range of methods, the overall code length is 
   - construct approximate wavefunctions with eigenvector continuation 
 
 ## Optional parameters
-For some parameters, the default values are used unless the user specifies those in the file, `optional_parameters.jl`.
+For some parameters, the default values are used unless the user specifies.
 See the [Optional parameters](parameters) page for more details.
 
 ## Issues/Pull requests
 
 NuclearToolkit.jl is designed to be an open-source software and to guarantee reproducibility and transparancy of the future works.
-Making issues and pull requests are fully welcome.
+Making issues and pull requests are fully welcome, you will be acknowledged or can be an author of future version of the paper.
+
