+++ src/chiEFTint/angmom_algebra.jl
@@ -96,7 +96,7 @@ end
 
 Function to carry out Talmi-Mochinsky transformation for NN interaction in HO space and to write out an sinput file.
 """
-function TMtrans(chiEFTobj,HOBs,to;calc_relcm=false,writesnt=true)
+function TMtrans(chiEFTobj::ChiralEFTobject,HOBs,to;calc_relcm=false,writesnt=true)
     V12ab = Vrel(chiEFTobj,to)
     params = chiEFTobj.params
     dLECs = chiEFTobj.LECs.dLECs; xr = chiEFTobj.xr; wr = chiEFTobj.wr;
@@ -131,7 +131,7 @@ function TMtrans(chiEFTobj,HOBs,to;calc_relcm=false,writesnt=true)
             println(io,"tza,  a, tzb,  b, tzc,  c, tzd,  d, J12,   <ab|v|cd>,   <ab|p_ip_j|cd>/mhw")
         elseif tbme_fmt=="snt"
             if length(target_nlj)!=0
-                nljdict = def_sps_snt(params.emax,params.target_nlj)[2]
+                nljdict = def_sps_snt(params.emax,target_nlj)[2]
             end
             write_spes(params,io,nljsnt,nofst,nTBME,nljdict)
         elseif tbme_fmt=="snt.bin"
@@ -181,7 +181,7 @@ function TMtrans(chiEFTobj,HOBs,to;calc_relcm=false,writesnt=true)
     end
 end
 
-function set_tbme(chiEFTobj,tbmes,ndim,izs,Jtot,vv,nljsnt,nljdict,tkeys,key6j,to;ofst=0)
+function set_tbme(chiEFTobj::ChiralEFTobject,tbmes,ndim,izs,Jtot,vv,nljsnt,nljdict,tkeys,key6j,to;ofst=0)
     dict6j = chiEFTobj.dict6j
     d6j_nabla = chiEFTobj.d6j_nabla
     target_nlj = chiEFTobj.params.target_nlj
@@ -243,8 +243,11 @@ function set_tbme(chiEFTobj,tbmes,ndim,izs,Jtot,vv,nljsnt,nljdict,tkeys,key6j,to
     return nothing
 end
 
+"""
+    red_nabla_l(n1,l1,n2,l2)
+returns ``b \\langle n_1,l_1|| \\nabla ||n_2,l_2 \\rangle`` in ``l``-reduced matrix element.
+"""
 function red_nabla_l(n1,l1,n2,l2)
-    # b*<n1,l1|| nabla ||n2,l2> in l-reduced m.e.
     if n1 == n2 && l1 == l2+1
         return -sqrt((l2 + 1.0)*(n2 + l2 + 1.5))
     elseif n1 == n2-1 && l1 == l2+1
@@ -257,18 +260,24 @@ function red_nabla_l(n1,l1,n2,l2)
         return 0.0
     end
 end
-
+"""
+    red_nabla_j(nlj1, nlj2)
+returns ```b \\langle j || \\nabla || j_2>```
+Note that ```l_1,l_2``` in `nlj1`&`nlj2` are not doubled.
+"""
 function red_nabla_j(nlj1, nlj2) 
-    # b*<j|| nabla ||j2> l1, l2 are not doubled 
     n1, l1, j1 = nlj1
     n2, l2, j2 = nlj2
     ret = (-1)^((3+2*l1+j2)//2) * sqrt(1.0*(j1+1)*(j2+1)) *
[…]
