+++ src/chiEFTint/struct_const_io.jl
@@ -22,7 +22,53 @@ const hc4 = hc^4
 const gA2 = gA^2
 const gA4 = gA^4
 const mpis = [139.5702,134.9766,139.5702]
-const fsalpha = 7.2973525693* 1.e-3 #fine structure const.   
+const fsalpha = 7.2973525693* 1.e-3 #fine structure const. 
+const l2l = [ wigner3j(Float64,l,2,l,0,0,0) for l=0:8]
+const l2lnd =[[ wigner3j(Float64,l1,2,l2,0,0,0) for l2=0:8] for l1=0:8]
+
+const nuclist = [
+     "H",  "He", "Li", "Be", "B",  "C",  "N",  "O",  "F", "Ne", "Na", "Mg", "Al", "Si", "P",  "S",  "Cl", "Ar", "K",  "Ca",
+    "Sc", "Ti", "V",  "Cr", "Mn", "Fe", "Co", "Ni", "Cu", "Zn", "Ga", "Ge", "As", "Se", "Br", "Kr", "Rb", "Sr", "Y",  "Zr",
+    "Nb", "Mo", "Tc", "Ru", "Rh", "Pd", "Ag", "Cd", "In", "Sn", "Sb", "Te", "I",  "Xe", "Cs", "Ba", "La", "Ce", "Pr", "Nd",
+    "Pm", "Sm", "Eu", "Gd", "Tb", "Dy", "Ho", "Er", "Tm", "Yb", "Lu", "Hf", "Ta", "W",  "Re", "Os", "Ir", "Pt", "Au", "Hg",
+    "Tl", "Pb", "Bi", "Po", "At", "Rn", "Fr", "Ra", "Ac", "Th", "Pa", "U",  "Np", "Pu", "Am", "Cm", "Bk", "Cf", "Es", "Fm",
+    "Md", "No", "Lr", "Rf", "Db", "Sg", "Bh", "Hs", "Mt", "Ds", "Rg", "Cn", "Nh", "Fl", "Mc", "Lv", "Ts", "Og" ]
+
+struct QLs
+    QL0s::Vector{Matrix{Float64}}
+    QWL2s::Vector{Matrix{Float64}}
+    QWL1s::Vector{Matrix{Float64}}
+    QWL0s::Vector{Matrix{Float64}}
+    QWlss::Vector{Matrix{Float64}}
+    ndQW1s::Vector{Vector{Matrix{Float64}}}
+    ndQW2s::Vector{Vector{Matrix{Float64}}}
+    QLdict::Vector{Dict{Float64,Float64}}
+end
+struct Fis_2n3n
+    F0s::Vector{Float64}
+    F1s::Vector{Float64}
+    F2s::Vector{Float64}
+    F3s::Vector{Float64}
+end
+struct wsyms_j1_1or2
+    cg1s::Matrix{Float64}
+    cg2s::Matrix{Float64}
+    d6_121::Array{Float64,3}
+    d6_21::Array{Float64,4}
+    d6_222::Array{Float64,3}
+    d9_12::Array{Float64,4}
+end
+struct util_2n3n
+    Fis::Fis_2n3n
+    QWs::QLs
+    wsyms::wsyms_j1_1or2
+end
+    
+struct LECs
+    vals::Vector{Float64}
+    idxs::Dict{String,Int64}
+    dLECs::Dict{String,Float64}
+end
 
 """
     chiEFTparams
@@ -73,6 +119,85 @@ mutable struct chiEFTparams
     kF::Float64
     BetaCM::Float64
 end
+
+"""
+    ChiralEFTobject
+# Fields
+- `params::chiEFTparams`
[…]
