+++ src/IMSRG.jl/commutator.jl
@@ -3,42 +3,41 @@
     
 overwrite `ret` operator by the commutator ``[X,Y]`` 
 """
-function OpCommutator!(X::Op,Y::Op,ret::Op,HFobj,Chan1b,Chan2bD,dictMono,dict6j,PandyaObj,to) where{Op <: Operator}
-    Chan2b = Chan2bD.Chan2b
-    E0 = ret.zerobody        
+function OpCommutator!(X::Op,Y::Op,ret::Op,HFobj::HamiltonianNormalOrdered,Chan1b::chan1b,Chan2bD::Chan2bD,dictMono,dict6j,PandyaObj::PandyaObject,to) where{Op <: Operator}
+    Chan2b = Chan2bD.Chan2b  
     if X.hermite && Y.hermite; ret.hermite=false; ret.antihermite=true
     elseif X.hermite && Y.antihermite; ret.hermite=true;ret.antihermite=false
     elseif X.antihermite && Y.hermite; ret.hermite=true;ret.antihermite=false;
     else ret.hermite = false;end
     ## zero-body pieces
     if !ret.antihermite
-        comm110ss!(X,Y,E0,HFobj)
-        comm220ss!(X,Y,E0,HFobj,Chan2b)
+        comm110ss!(X,Y,ret,HFobj)
+        comm220ss!(X,Y,ret,HFobj,Chan2b)
     end    
     ## one-body pieces
     @timeit to "comm111" comm111ss!(X,Y,ret)
-    @timeit to "comm121" comm121ss!(X,Y,ret,HFobj,Chan1b,Chan2b,dictMono,PandyaObj)
+    @timeit to "comm121" comm121ss!(X,Y,ret,HFobj,Chan1b,dictMono,PandyaObj)
     @timeit to "comm221" comm221ss!(X,Y,ret,HFobj,Chan1b,Chan2bD,PandyaObj)
     ## two-body pieces
-    @timeit to "comm122" comm122ss!(X,Y,ret,HFobj,Chan1b,Chan2b,PandyaObj,to)
+    @timeit to "comm122" comm122ss!(X,Y,ret,Chan2b,PandyaObj,to)
     @timeit to "comm222pphh" comm222_pphh_ss!(X,Y,ret,HFobj,Chan2bD,PandyaObj,to)
     @timeit to "comm222ph" comm222ph_ss!(X,Y,ret,HFobj,Chan2bD,dict6j,PandyaObj,to)
     return nothing
 end
 
 """
-    BCH_Product(X,Y,Z,tmpOp,Nested,ncomm,norms,Chan1b,Chan2bD,HFobj,dictMono,dict6j,PandyaObj,to;tol=1.e-4)
+    BCH_Product(X::Op,Y::Op,Z::Op,tmpOp::Op,Nested::Op,ncomm::Vector{Int64},norms::Vector{Float64},Chan1b::chan1b,Chan2bD::Chan2bD,HFobj::HamiltonianNormalOrdered,
+dictMono,dict6j,PandyaObj::PandyaObject,to;tol=1.e-4) where Op <:Operator
 
-returns ``Z``  to satisfy: ``e^Z = e^X e^Y``  
+returns ``Z``  to satisfy: ``e^Z = e^X e^Y``.
 
 ``Z`` is calculated with Baker–Campbell–Hausdorff (BCH) formula:
-
 ``Z = X + Y + 1/2[X, Y]  + 1/12 [X,[X,Y]] + 1/12 [Y,[Y,X]] -1/24 [Y,[X,[X,Y]]] -1/720 [Y,[Y,[Y,[Y,X]]]] -1/720 [X,[X,[X,[X,Y]]]] +...``
 
[…]
