+++ src/chiEFTint/angmom_algebra.jl
@@ -92,11 +92,11 @@ function s_wigner9j(j1,j3,j4,j6,j7,j8,j9)
 end
 
 """
-    TMtrans(chiEFTobj,HOBs,to;calc_relcm=false,writesnt=true)
+    TMtrans(chiEFTobj::ChiralEFTobject,HOBs::Dict{Int64, Dict{Int64,Float64}},to;calc_relcm=false,writesnt=true)
 
 Function to carry out Talmi-Mochinsky transformation for NN interaction in HO space and to write out an sinput file.
 """
-function TMtrans(chiEFTobj::ChiralEFTobject,HOBs,to;calc_relcm=false,writesnt=true)
+function TMtrans(chiEFTobj::ChiralEFTobject,HOBs::Dict{Int64, Dict{Int64,Float64}},to;calc_relcm=false,writesnt=true)
     V12ab = Vrel(chiEFTobj,to)
     params = chiEFTobj.params
     dLECs = chiEFTobj.LECs.dLECs; xr = chiEFTobj.xr; wr = chiEFTobj.wr;
@@ -145,7 +145,6 @@ function TMtrans(chiEFTobj::ChiralEFTobject,HOBs,to;calc_relcm=false,writesnt=tr
     tkeys = [zeros(Int64,4) for i=1:4]
     key6j = zeros(Int64,5)
     t5vs=[zeros(Int64,5) for i=1:nthreads()]   
-    ttt = [deepcopy(HOBs) for i=1:nthreads()]   
     @inbounds for ich in eachindex(infos) 
         izz,ip,Jtot,ndim=infos[ich]
         pnrank = Int(div(izz,2))+2
@@ -153,7 +152,6 @@ function TMtrans(chiEFTobj::ChiralEFTobject,HOBs,to;calc_relcm=false,writesnt=tr
         vv = zeros(Float64,ndim,ndim)
         @timeit to "vtrans" @inbounds @threads for i = 1:ndim
             t5v = t5vs[threadid()]
-            HOBs = ttt[threadid()]
             iza,ia,izb,ib = izs[i]
             @inbounds for j = 1:i
                 izc,ic,izd,id= izs[j]
@@ -262,24 +260,22 @@ function red_nabla_l(n1,l1,n2,l2)
 end
 """
     red_nabla_j(nlj1, nlj2)
-returns ```b \\langle j || \\nabla || j_2>```
-Note that ```l_1,l_2``` in `nlj1`&`nlj2` are not doubled.
+returns ``b \\langle j || \\nabla || j_2\\rangle``
+Note that ``l_1,l_2`` in `nlj1`&`nlj2` are not doubled.
 """
 function red_nabla_j(nlj1, nlj2) 
-    n1, l1, j1 = nlj1
-    n2, l2, j2 = nlj2
+    n1, l1, j1 = nlj1; n2, l2, j2 = nlj2
     ret = (-1)^((3+2*l1+j2)//2) * sqrt(1.0*(j1+1)*(j2+1)) *
         wigner6j(Float64,j1//2, 1, j2//2, l2, 1//2, l1) * red_nabla_l(n1,l1,n2,l2)
     return ret
 end
 """
     red_nabla_j(nlj1,nlj2,d6j,key6j) 
-returns ```b \\langle j || \\nabla || j_2>```
[…]

[…]
