### 2022-09-26 #4694 @SotaYoshida
@mdavezac Apologies for the delay.

I have made dozens of modifications on the package according to your valuable comment, and then tagged the latest one as v2.0.1.
(For the above command , I didn't understand how to interact with the bot. Please do not care.)

Here is the summary and reply to your comment:

1. For the documantation

    I have made new pages, [HowToUse](https://sotayoshida.github.io/NuclearToolkit.jl/dev/howtouse/) and [FileFormat](https://sotayoshida.github.io/NuclearToolkit.jl/dev/fileformat/) to show example usage of the main APIs and to explain what input/output files are, on the [document](https://sotayoshida.github.io/NuclearToolkit.jl/dev/).

2. For test codes
    >there is no example or test for MPI (the code claims to work on supercomputers)

    Absolutely. An example code and a test are added to the document and CI test.
    I am not fully aware of how much is required in terms of unit testing, but I have added a test code to validate the outputs of many-body methods with a specific parameter set. I think this somewhat improved the mitigation of destructive changes during the code development.

3. > the API documentation is too often sparse, especially when one considers some functions listed there take 10-20 arguments with no obvious typing. Since this package is meant as a toolkit for the community, I would recommend (though not require) investing some time into it. In this respect, unit-tests demonstrating usage and expected outputs would have been valuable.

    Yes. First, I have modified main APIs, especially `make_chiEFTint()`, and structure of functions therein, and the package is now available to run outside the repository. [#27](https://github.com/SotaYoshida/NuclearToolkit.jl/issues/27)
     Second, I have added detailed docstrings on the main APIs and homemade (mutable) structs, and added annotations on the structs so that a user can trace the workflow of the code by e.g. GUI editor like VScode.

4.  I fixed the typos in the paper

I know you are very busy, but I hope you could give us another review. Thank you so much for your time.
