### 2022-09-26 #4694 @rkurchin
@SotaYoshida thanks for all your work on these changes! Just so you know, adding the version tag via the bot is something only I as the editor have the permission to do. I usually wait until the end of the review to do that so it can just be set when it's finalized. So don't worry about it!
