### 2022-10-02 SotaYoshida/NuclearToolkit.jl#31 @villaa
[JOSS Review] use doi's and links for arXiv references
Several references in the reference list do not list links or DOIs, but there are papers on arXiv which gives links/DOIs. I suggest using the style found in the paper [here](https://joss.theoj.org/papers/10.21105/joss.03725) for those references.

The references are:

Bezanson, J., Karpinski, S., Shah, V. B., & Edelman, A. (2012). Julia: A fast dynamic45
language for technical computing. arXiv Preprint arXiv:1209.5145.

Johnson, C. W., Ormand, W. E., McElvain, K. S., & Shan, H. (2018). BIGSTICK: A flexible64
configuration-interaction shell-model code.

Shimizu, N. (2013). Nuclear shell-model code for massive parallel computation, ”KSHELL”.

Yoshida, S. (2022). Exploring medium mass nuclei using effective chiral nucleon-nucleon86
interactions. arXiv Preprint arXiv:2208.02464.
