### 2022-10-02 #4694 @villaa
@SotaYoshida I started reviewing, I think the paper is well written I submitted one [issue ](https://github.com/SotaYoshida/NuclearToolkit.jl/issues/31) on GitHub relating to reference links and DOIs for arXiv-only references. I generally tag the issues related to my review with [JOSS Review].

I will continue reviewing the software!
