+++ paper/paper.bib
@@ -3,14 +3,18 @@
   title={{J}ulia: A fast dynamic language for technical computing},
   author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
   journal={arXiv preprint arXiv:1209.5145},
-  year={2012}
+  year={2012},
+  doi = {10.48550/ARXIV.1209.5145},  
+  url = {https://arxiv.org/abs/1209.5145},
 }
 
 @article{SY2n3n,
   title={Exploring medium mass nuclei using effective chiral nucleon-nucleon interactions},
   author={Yoshida, Sota},
   journal={arXiv preprint arXiv:2208.02464},
-  year={2022}
+  year={2022},
+  doi = {10.48550/ARXIV.2208.02464},
+  url = {https://arxiv.org/abs/2208.02464},
 }
 
 @article{BrownRev,
@@ -190,7 +194,9 @@ keywords = {Shell model, Configuration interaction, Many-body}
 Author = {Calvin W. Johnson and W. Erich Ormand and Kenneth S. McElvain and Hongzhang Shan},
 Title = {BIGSTICK: A flexible configuration-interaction shell-model code},
 Year = {2018},
-Eprint = {arXiv:1801.08432},
+journal={arXiv preprint arXiv:1801.08432},
+doi = {10.48550/ARXIV.1801.08432},
+url = {https://arxiv.org/abs/1801.08432}
 }
 @article{NuShellX,
 title = "The Shell-Model Code NuShellX@MSU",
@@ -206,7 +212,9 @@ doi = "https://doi.org/10.1016/j.nds.2014.07.022"
 Author = {Noritaka Shimizu},
 Title = {Nuclear shell-model code for massive parallel computation, "KSHELL"},
 Year = {2013},
-Eprint = {arXiv:1310.5431}
+journal = {arXiv preprint arXiv:1310.5431}
+doi = {10.48550/ARXIV.1310.5431},
+url = {https://arxiv.org/abs/1310.5431},
 }
 @article{KSHELL2,
 title = "Thick-restart block Lanczos method for large-scale shell-model calculations",
