+++ paper/paper.bib
@@ -5,7 +5,7 @@
   journal={arXiv preprint arXiv:1209.5145},
   year={2012},
   doi = {10.48550/ARXIV.1209.5145},  
-  url = {https://arxiv.org/abs/1209.5145},
+  url = {https://arxiv.org/abs/1209.5145}
 }
 
 @article{SY2n3n,
@@ -14,7 +14,7 @@
   journal={arXiv preprint arXiv:2208.02464},
   year={2022},
   doi = {10.48550/ARXIV.2208.02464},
-  url = {https://arxiv.org/abs/2208.02464},
+  url = {https://arxiv.org/abs/2208.02464}
 }
 
 @article{BrownRev,
@@ -212,7 +212,7 @@ doi = "https://doi.org/10.1016/j.nds.2014.07.022"
 Author = {Noritaka Shimizu},
 Title = {Nuclear shell-model code for massive parallel computation, "KSHELL"},
 Year = {2013},
-journal = {arXiv preprint arXiv:1310.5431}
+journal = {arXiv preprint arXiv:1310.5431},
 doi = {10.48550/ARXIV.1310.5431},
 url = {https://arxiv.org/abs/1310.5431},
 }
