+++ paper/paper.bib
@@ -4,7 +4,7 @@
   author={Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B and Edelman, Alan},
   journal={arXiv preprint arXiv:1209.5145},
   year={2012},
-  doi = {10.48550/ARXIV.1209.5145},  
+  doi = {10.48550/arXiv.1209.5145},  
   url = {https://arxiv.org/abs/1209.5145}
 }
 
@@ -13,7 +13,7 @@
   author={Yoshida, Sota},
   journal={arXiv preprint arXiv:2208.02464},
   year={2022},
-  doi = {10.48550/ARXIV.2208.02464},
+  doi = {10.48550/arXiv.2208.02464},
   url = {https://arxiv.org/abs/2208.02464}
 }
 
@@ -195,7 +195,7 @@ Author = {Calvin W. Johnson and W. Erich Ormand and Kenneth S. McElvain and Hong
 Title = {BIGSTICK: A flexible configuration-interaction shell-model code},
 Year = {2018},
 journal={arXiv preprint arXiv:1801.08432},
-doi = {10.48550/ARXIV.1801.08432},
+doi = {10.48550/arXiv.1801.08432},
 url = {https://arxiv.org/abs/1801.08432}
 }
 @article{NuShellX,
@@ -213,7 +213,7 @@ Author = {Noritaka Shimizu},
 Title = {Nuclear shell-model code for massive parallel computation, "KSHELL"},
 Year = {2013},
 journal = {arXiv preprint arXiv:1310.5431},
-doi = {10.48550/ARXIV.1310.5431},
+doi = {10.48550/arXiv.1310.5431},
 url = {https://arxiv.org/abs/1310.5431},
 }
 @article{KSHELL2,
