+++ paper/paper.bib
@@ -177,6 +177,7 @@ url={https://doi.org/10.1140/epja/i2003-10096-0}
    adsurl = {https://ui.adsabs.harvard.edu/abs/1999AcPPB..30..705C},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 @article{BIGSTICK,
 title = {Factorization in large-scale many-body calculations},
 journal = {Computer Physics Communications},
@@ -185,16 +186,15 @@ number = {12},
 pages = {2761-2774},
 year = {2013},
 issn = {0010-4655},
-doi = {https://doi.org/10.1016/j.cpc.2013.07.022},
-url = {https://www.sciencedirect.com/science/article/pii/S0010465513002518},
 author = {Calvin W. Johnson and W. Erich Ormand and Plamen G. Krastev},
-keywords = {Shell model, Configuration interaction, Many-body}
+doi = {10.1016/j.cpc.2013.07.022},
+url = {https://www.sciencedirect.com/science/article/pii/S0010465513002518},
 }
 @misc{BIGSTICK2,
-Author = {Calvin W. Johnson and W. Erich Ormand and Kenneth S. McElvain and Hongzhang Shan},
 Title = {BIGSTICK: A flexible configuration-interaction shell-model code},
-Year = {2018},
-journal={arXiv preprint arXiv:1801.08432},
+year = {2018},
+journal = {arXiv preprint arXiv:1801.08432},
+author = {Calvin W. Johnson and W. Erich Ormand and Kenneth S. McElvain and Hongzhang Shan},
 doi = {10.48550/arXiv.1801.08432},
 url = {https://arxiv.org/abs/1801.08432}
 }
