### 2022-10-04 SotaYoshida/NuclearToolkit.jl#31 @SotaYoshida
Thanks again. I hope the DOIs are now properly added.
https://github.com/openjournals/joss-reviews/issues/4694#issuecomment-1266316048
