+++ src/chiEFTint/pionexchange.jl
@@ -18,34 +18,34 @@ function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
     mpipm = mpis[1]; mpipm2 = mpipm^2
     for pnrank =1:3
         tdict = dict_pwch[pnrank]
-        MN = Ms[pnrank];dwn = 1.0/MN;sq_dwn=dwn^2
-        fff = pi / ((2*pi)^3 * MN^2)
+        MN = Ms[pnrank];dwn = 1.0/MN
+        fff = pi / ((2*pi)^3) * dwn^2
         coeff = -(MN*gA/(2*Fpi))^2
         itt = 2 *(pnrank -2)
         tllsj[1] = itt
         @inbounds for J=0:jmax
             lsj = lsjs[J+1]
             f_idx = 6
-            if J==0; f_idx = 3;end                
+            if J==0; f_idx = 3;end         
             @inbounds for i= 1:n_mesh
                 x = xr[i];xdwn = x * dwn;sq_xdwn= xdwn^2
                 ex = sqrt(1.0+sq_xdwn)
                 @inbounds for j = 1:n_mesh
                     y = xr[j]; ydwn = y*dwn;sq_ydwn= ydwn^2
                     ey = sqrt(1.0+sq_ydwn)
-                    nfac = 1.0/(x* y* sq_dwn)
+                    nfac = 1.0 / (xdwn * ydwn)
                     ree = 1.0/sqrt(ex*ey) * freg(x,y,4)
                     f_sq!(opfs,xdwn,ydwn)
                     if pnrank != 2 ## pp/nn
-                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,nfac,ts,ws,tVs,QLdict)
+                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,ts,ws,tVs,QLdict)
                     else ## pn
-                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,nfac,ts,ws,tVs,QLdict)
-                        cib_lsj_opep(opfs,x,y,mpipm2,2,J,pnrank,nfac,ts,ws,tVs,QLdict;additive=true)
+                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,ts,ws,tVs,QLdict)
+                        cib_lsj_opep(opfs,x,y,mpipm2,2,J,pnrank,ts,ws,tVs,QLdict;additive=true)
                         #if pigamma
                         #    cib_lsj_opep(opfs,x,y,mpipm2,2,J,pnrank,nfac,ts,ws,tVs,pigamma;additive=true)
                         #end
                     end
-                    t_fc = hc3 * fff * coeff * ree
+                    t_fc = fff * coeff * ree *hc3 * nfac                   
                     @inbounds for idx = 1:f_idx
                         @views tllsj[2:5] .= lsj[idx]
                         tl,tlp,tS,tJ = lsj[idx] 
@@ -53,7 +53,7 @@ function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
                         V12idx = get(tdict,tllsj,-1)
                         if V12idx == -1;continue;end                        
                         tfac = tVs[idx] * t_fc
-                        V12mom[V12idx][i,j] += tfac
[…]
