+++ src/chiEFTint/main_chiEFTint.jl
@@ -26,10 +26,19 @@ function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="
     @timeit to "deutron" BE_d = Calc_Deuteron(chiEFTobj,to)
     @timeit to "renorm." SRG(chiEFTobj,to)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc) 
+
+    # QLdict = chiEFTobj.util_2n3n.QWs.QLdict
+    # println("keys QLdict[J=2] ",keys(QLdict[3]))
+
+    target_LSJ = [[0,0,1,1],[1,1,1,0],[1,1,0,1],[1,1,1,1],[0,0,0,0],[0,2,1,1],[3,3,1,3]];write_onshell_vmom(chiEFTobj,2,target_LSJ;label="pn")
+
     if do2n3ncalib #calibrate 2n3n LECs by HFMBPT
+        ## not yet updated to seperate 2n3n from NN
         caliblating_2n3nLECs_byHFMBPT(itnum,optimizer,MPIcomm,chiEFTobj,OPTobj,d9j,HOBs,nucs,HFdata,to,io;Operators=Operators)        
     else # write out snt/snt.bin file
-        add2n3n(chiEFTobj,to)
+        calc_vmom_3nf(chiEFTobj,1,to)
+        if chiEFTobj.params.calc_EperA; calc_nuclearmatter_in_momspace(chiEFTobj,to);end
+        #add2n3n(chiEFTobj,to)
         @timeit to "Vtrans" dicts_tbme = TMtrans(chiEFTobj,HOBs,to;writesnt=writesnt)
     end
     if io != stdout; close(io);end
@@ -50,7 +59,7 @@ function construct_chiEFTobj(do2n3ncalib,itnum,optimizer,MPIcomm,io,to;fn_params
     params = init_chiEFTparams(;io=io,fn_params=fn_params)
     #Next, prepare momentum/integral mesh, arrays, etc.
     ## Prep WignerSymbols
-    dict6j,d6j_nabla,d6j_int = PreCalc6j(params.emax,!do2n3ncalib)
+    dict6j,d6j_nabla,d6j_int = PreCalc6j(params.emax)
     ## prep. momentum mesh
     xr_fm,wr = Gauss_Legendre(0.0,params.pmax_fm,params.n_mesh); xr = xr_fm .* hc
     pw_channels,dict_pwch,arr_pwch = prepare_2b_pw_states(;io=io)    
@@ -76,7 +85,7 @@ function construct_chiEFTobj(do2n3ncalib,itnum,optimizer,MPIcomm,io,to;fn_params
     infos,izs_ab,nTBME = make_sp_state(params;io=io)
     println(io,"# of channels 2bstate ",length(infos)," #TBME = $nTBME")
     ## prep. integrals for 2n3n
-    util_2n3n = prep_integrals_for2n3n(params,xr,ts,ws)
+    @timeit to "util2n3n" util_2n3n = prep_integrals_for2n3n(params,xr,ts,ws)
     ### specify low-energy constants (LECs)
     LECs = read_LECs(params.pottype)
     ## 9j&6j symbols for 2n (2n3n) interaction
@@ -87,8 +96,8 @@ function construct_chiEFTobj(do2n3ncalib,itnum,optimizer,MPIcomm,io,to;fn_params
                             infos,izs_ab,nTBME,util_2n3n,LECs,X9,U6,
                             V12mom,V12mom_2n3n,pw_channels,dict_pwch,arr_pwch)
     # make Opt stuff    
[…]
