+++ docs/src/fileformat.md
@@ -36,3 +36,12 @@ Here is an example of w.snt. The comments are not written in the outputs of Nucl
 1 1 1 1 0 -2.18450
 1 1 1 1 2 -0.06650
 ```
+
+For free space interactions, output of `chiEFTint`, have V2n3n and Vpp terms in addition to the VNN term:
+```
+...
+! i j k l J <i,j| VNN | k,l>_J <i,j| V2n3n | k,l>_J <i,j| Vpp | k,l>_J
+  1    1    1    1     0     -6.4492081719      0.0000000000   -0.0000000000e+00
+  1    2    1    2     0     -1.9671496992      0.0000000000    5.0000000000e-01
+...
+```
