+++ docs/src/howtouse.md
@@ -9,6 +9,7 @@ A sample of `optional_parameters.jl` is available on the [repository](https://gi
 
 The main API is `make_chiEFTint()` and the sample code is given below.
 ```julia
+using NuclearToolkit
 #1). generate NN interaction file in snt or snt.bin format
 make_chiEFTint()
 
@@ -48,6 +49,7 @@ You can evaluate ground state properties by the so-called Hartree-Fock many-body
 You must specify target nuclei, snt file (generated by chiEFTint), harmonic oscillator parameter `hw`, model space size `emax`.
 
 ```julia
+using NuclearToolkit
 hw = 20; emax=4
 sntf = "path_to_your.snt"
 #1). you can evaluate HF-MBPT(3) estimate of g.s. energy of the target nuclus
@@ -67,6 +69,7 @@ hf_main(nucs,sntf,hw,emax;Operators=["Rp2"])
 
 One can perform in-medium similarity renormalization group (IM-SRG) calculations by specifing the optional argument `doIMSRG=true`:
 ```julia
+using NuclearToolkit
 hw = 20; emax=4
 nucs = ["He4","O16"]
 sntf = "path_to_your.snt"
@@ -82,6 +85,7 @@ You can also perform valence-space IMSRG (VS-IMSRG) decoupling to derive shell-m
 The API is similar to the IMSRG case, but you must specify `core`, `ref`, and `vspace`.
 
 ```julia
+using NuclearToolkit
 hw = 20; emax=4
 core = "O16"; vspace="sd-shell", ref="core"
 nucs = ["Mg24"]
@@ -101,6 +105,7 @@ You need to specify the effective interaction (in snt format), number of eigenva
 to perform shell-model calculations.
 
 ```julia
+using NuclearToolkit
 hw = 20; emax=4;
 vs_sntf = "vsimsrg_sd-shell_coreO16refO16_O16_hw20e4_Delta0.0.snt"
 n_eigen=10
