+++ test/mpirun_test.jl
@@ -1,4 +1,3 @@
-#import Pkg; Pkg.add("NuclearToolkit")
 using NuclearToolkit
 using Test
 
