+++ docs/src/contributing.md
@@ -1,5 +1,7 @@
 # Contributing to NuclearToolkit.jl
 
+Thank you for considering contributing to this package.
+
 Feedbacks and contributions to NuclearToolkit.jl are very welcome.
 These can be:
 - bug report
@@ -8,26 +10,27 @@ These can be:
 - feature request
 - etc.
 
-For these contributions, it would be nice to let you know a basic policy (the workflow of code development, and LICENSE, shown below) in this package.
-It should be noted that the package has been developed by a single author (@SotaYoshida) so far, and thereby the followings are just the author's policy. Comments on development policy are also welcome.
+For these contributions, it would be nice to let you know a basic policy (the workflow of code development and LICENSE, shown below) in this package.
+Note that the package has been developed by a single author (@SotaYoshida) so far, and thereby the followings are just the author's policy. Comments on the development policy itself are also welcome.
 
 
 ## Workfklow of code development
 
 We use the GitHub to host the package, to track issues/pull requests.
 
-We use the Julia package, Documenter.jl, to build the document.
-About 
-- release
-- 
+The document is built using [Documenter.jl](https://juliadocs.github.io/Documenter.jl/stable/),
+which is a package for building documentation from docstrings and markdown files.
+It is automized to build and deploy the document by GitHub Actions.
 
-workflow (which branches to submit PR to, etc)
+A typical workfklow is the following:
 
-GitHub Actions (.github/workflow/) run the test code
+1. clone or folk the repository and modify the code and/or document locally
 
-.github/workflows/CI.yml
+2. propose modifications through a pull request (PR) to **'dev' branch**
 
-When you want to add a new API to the package, which can be called in Julia REPL, you need to export it `src/NuclearToolkit.jl`.
+3. if the PR passes the automated tests, it will be merged 
+
+4. At some point, we tag and release, which is done automatically by instructing the JuliaRegistrator bot to do so.
 
 
 ### Automated tests on GitHub Actions
@@ -35,37 +38,47 @@ When you want to add a new API to the package, which can be called in Julia REPL
 We use GitHub Actions to run the test codes and to build/deploy the document.
 When some changes are submitted through a pull request, the test codes are run to check that the changes are not destructive.
 
-The test codes and yml files specifying them are in `test` in the repository. They checks
+The test jobs are specified in yml files like `.github/workflows/CI.yml` and one can find the test code in `test/` of the repository.
+
+They checks
 - Generating a nucleon-nucleon potential, i.e., input for nuclear many-body methods
-- HFMBPT results with it
-- IMSRG/VSIMSRG results with it
-- shell-model results with the derived effective interaction
+- HFMBPT results with the NN-potential
+- IMSRG/VSIMSRG results the NN-potential
+- shell-model results with the derived effective interaction by VS-IMSRG
 - MPI and sampling stuffs are runnable
 
-If you submit a major change to the code and it brings the changes in the numbers of the result by many-body methods,
+If you submit a major change to the code and it brings the changes in the results by many-body methods,
 it is recommended to contact the main author(s). Otherwise, we appreciate it if you could propose updated test codes.
 
 
-## Report a bug by opening a new issue
+## Reporting bugs by opening a new issue
 
 Thank you so much for considering to report bugs!
-When you report a bug in the code, please open an issue and make sure to include any information necessary for us to reproduce the bug.
+When you report a bug in the code, please open an new issue from [Here](https://github.com/SotaYoshida/NuclearToolkit.jl/issues).
+Please make sure to include any information necessary for us to reproduce the errors. Thanks!
 
+## Propose modifications through Pull Requests (PRs)
 
[…]
