### 2022-10-05 f1ff4f6
verbose option changed and minor fix of docs
