### 2022-10-05 SotaYoshida/NuclearToolkit.jl#31 @villaa
> Thanks again. I hope the DOIs are now properly added. [openjournals/joss-reviews#4694 (comment)](https://github.com/openjournals/joss-reviews/issues/4694#issuecomment-1266316048)

Yes, they look good now. I think this issue can be closed--I don't know if you want to do a pull request. I often try to make pull requests and issues 1-to-1 thanks to a suggestion from my first JOSS review.

If you want to do that: make a pull request from the branch that you have this fix on and include this issue as  being addressed by that request. Then they are linked and you can close both at the same time.
