### 2022-10-06 11b01f2
#31 using DOI and minor fix done
