### 2022-10-10 SotaYoshida/NuclearToolkit.jl#38 @mdavezac
[JOSS REVIEW] IO errors when running with MPI
When running with mpi and multiple processors, I get the following errors:

1. Anything printed to the stdout is printed identically multiple times
1. At the very end, after printing the results, an error occurs:
   > SystemError: opening file "flowOmega/Omega_21970O16_1.bin": No such file or directory

The first error certainly is a result of printing outputs from all processes, rather than from the main process only. Perhaps some `verbose` variable can be set to `false` for other processes other than main?

I guess the second error also comes from trying to write from all processes. However, if the information written to the binary files is not the same on all procs, it might be more complicated to correct.

Part of openjournals/joss-reviews#4694.
