### 2022-10-10 SotaYoshida/NuclearToolkit.jl#38 @mdavezac
For the record, 1. is an annoyance but it shouldn't hold back publication. However, I would think 2 affects the functionality and usability of the code and should be required to claim that the code works with MPI.
