### 2022-10-11 SotaYoshida/NuclearToolkit.jl#38 @SotaYoshida
@mdavezac Thank you again!

In summary, this issue may not be on the code, but the document. Let me try to explain.

From the error message,
> SystemError: opening file "flowOmega/Omega_21970O16_1.bin": No such file or directory

you may have run in-medium SRG (IM-SRG) calculations with MPI.
The error is due to the [function](https://github.com/SotaYoshida/NuclearToolkit.jl/blob/main/src/IMSRG.jl/imsrg_util.jl#L638) to clean previous temporary files `flowOmega/*`, which are generated through solving matrix ODEs repeatedly.

>For the record, 1. is an annoyance but it shouldn't hold back publication. However, I would think 2 affects the functionality and usability of the code and should be required to claim that the code works with MPI.

I am afraid to say... in the first place, I am not claiming one can benefit from multi-node parallelization (using e.g. MPI.jl) for many-body methods in the package.
In this package, the MPI is used explicitly only for MCMC samplings using HF-MBPT (one many-body method) and affine-invariant MCMC, which is a kind of ensemble method that uses the results of independent runs to make MCMC sampling efficient.

My intent
- in the paper
    > ... works on both laptops and supercomputers in almost the same way.
- in the document
    >  The code can be easily executed on either a laptop or a supercomputer.

was... one can easily run the code on different environments due to Julia's Pkg manager and portability, and I did not mean to say one can benefit from MPI elsewhere.

Sorry for the confusion.
I will improve the document for both #38 and #39.
