### 2022-10-11 SotaYoshida/NuclearToolkit.jl#38 @mdavezac
Okay, thanks for clearing that up. I misunderstood the claim "works on supercomputers" as "works with MPI". To be fair, that might be a common misunderstanding :). Perhaps it would be a good idea to document where MPI is expected to work explicitly.

A suggestion: Have you tested this on Windows? Perhaps the same idea can be conveyed by saying "works on a variety of environments, including Mac, linux, and Windows"? This shouldn't result in fewer understanding about hardware.
