+++ src/chiEFTint/main_chiEFTint.jl
@@ -1,5 +1,5 @@
 """
-    make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[])
+    make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[],fn_params="optional_parameters.jl",write_vmom=false)
 
 The interface function in chiEFTint.
 This generates NN-potential in momentum space and then transforms it in HO basis to give inputs for many-body calculations.
@@ -13,8 +13,10 @@ The function is exported and can be simply called make_chiEFTint() in your scrip
 - `optimizer::String` method for LECs calibration. "MCMC","LHS","BayesOpt" are available
 - `MPIcomm::Bool`, to carry out LECs sampling with HF-MBPT and affine inveriant MCMC
 - `Operators::Vector{String}` specifies operators you need to use in LECs calibrations
+- `fn_params::String` path to file specifying the optional parameters
+- `write_vmom::Bool` to write out in vmom partial wave channels
 """
-function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[],fn_params="optional_parameters.jl")
+function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="",MPIcomm=false,corenuc="",ref="nucl",Operators=[],fn_params="optional_parameters.jl",write_vmom=false)
     to = TimerOutput()
     do2n3ncalib=false
     if (optimizer!="" && nucs != []) || MPIcomm
@@ -23,21 +25,21 @@ function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="
     io = select_io(MPIcomm,optimizer,nucs)
     @timeit to "prep." chiEFTobj,OPTobj,d9j,HOBs = construct_chiEFTobj(do2n3ncalib,itnum,optimizer,MPIcomm,io,to;fn_params)
     @timeit to "NNcalc" calcualte_NNpot_in_momentumspace(chiEFTobj,to)
-    @timeit to "deutron" BE_d = Calc_Deuteron(chiEFTobj,to)
+    @timeit to "deutron" BE_d = Calc_Deuteron(chiEFTobj,to;io=io)
     @timeit to "renorm." SRG(chiEFTobj,to)
     HFdata = prepHFdata(nucs,ref,["E"],corenuc) 
 
-    # QLdict = chiEFTobj.util_2n3n.QWs.QLdict
-    # println("keys QLdict[J=2] ",keys(QLdict[3]))
-
-    target_LSJ = [[0,0,1,1],[1,1,1,0],[1,1,0,1],[1,1,1,1],[0,0,0,0],[0,2,1,1],[3,3,1,3]];write_onshell_vmom(chiEFTobj,2,target_LSJ;label="pn")
+    if write_vmom
+        target_LSJ = [[0,0,1,1],[1,1,1,0],[1,1,0,1],[1,1,1,1],[0,0,0,0],[0,2,1,1],[3,3,1,3]]
+        write_onshell_vmom(chiEFTobj,2,target_LSJ;label="pn")
+    end
 
     if do2n3ncalib #calibrate 2n3n LECs by HFMBPT
         ## not yet updated to seperate 2n3n from NN
[…]
