+++ src/chiEFTint/pionexchange.jl
@@ -1,17 +1,16 @@
 """
-    OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
+    OPEP(chiEFTobj,to;pigamma=false)
 
 calc. One-pion exchange potential in the momentum-space
 
 Reference: R. Machleidt, Phys. Rev. C 63 024001 (2001).
 """
-function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
+function OPEP(chiEFTobj,to;pigamma=false)
     ts = chiEFTobj.ts; ws = chiEFTobj.ws; xr = chiEFTobj.xr; V12mom = chiEFTobj.V12mom
     dict_pwch = chiEFTobj.dict_pwch; 
     lsjs = chiEFTobj.lsjs; tllsj = chiEFTobj.tllsj
     opfs = chiEFTobj.opfs; QLdict = chiEFTobj.util_2n3n.QWs.QLdict
     n_mesh = chiEFTobj.params.n_mesh
-    hc3 = hc^3
     tVs = zeros(Float64,6)
     opfs = zeros(Float64,8)
     mpi0 = mpis[2]; mpi02 = mpi0^2
@@ -19,14 +18,13 @@ function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
     for pnrank =1:3
         tdict = dict_pwch[pnrank]
         MN = Ms[pnrank];dwn = 1.0/MN
-        fff = pi / ((2*pi)^3) * dwn^2
-        coeff = -(MN*gA/(2*Fpi))^2
+        fff = - gA2 *  pi / ((2*Fpi)^2 * (2*pi)^3)
         itt = 2 *(pnrank -2)
         tllsj[1] = itt
         @inbounds for J=0:jmax
             lsj = lsjs[J+1]
             f_idx = 6
-            if J==0; f_idx = 3;end         
+            if J==0; f_idx = 3;end 
             @inbounds for i= 1:n_mesh
                 x = xr[i];xdwn = x * dwn;sq_xdwn= xdwn^2
                 ex = sqrt(1.0+sq_xdwn)
@@ -34,26 +32,25 @@ function OPEP(chiEFTobj,to;pigamma=true,debugmode=false)
                     y = xr[j]; ydwn = y*dwn;sq_ydwn= ydwn^2
                     ey = sqrt(1.0+sq_ydwn)
                     nfac = 1.0 / (xdwn * ydwn)
-                    ree = 1.0/sqrt(ex*ey) * freg(x,y,4)
+                    ree =  1.0 /sqrt(ex*ey) *freg(x,y,4)
                     f_sq!(opfs,xdwn,ydwn)
-                    if pnrank != 2 ## pp/nn
-                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,ts,ws,tVs,QLdict)
-                    else ## pn
-                        cib_lsj_opep(opfs,x,y,mpi02,1,J,pnrank,ts,ws,tVs,QLdict)
-                        cib_lsj_opep(opfs,x,y,mpipm2,2,J,pnrank,ts,ws,tVs,QLdict;additive=true)
-                        #if pigamma
-                        #    cib_lsj_opep(opfs,x,y,mpipm2,2,J,pnrank,nfac,ts,ws,tVs,pigamma;additive=true)
-                        #end
+                    t_fc = fff * ree *hc3 * nfac
+                    tVs .= 0.0
[…]
