+++ .github/workflows/CI.yml
@@ -21,7 +21,7 @@ jobs:
         include:
           - {version: '1.7',   os: ubuntu-latest,  payload: serial}
           - {version: '1.7',   os: macOS-latest,   payload: serial}
-	        - {version: '1.7',   os: Windows,   payload: serial}
+	        - {version: '1.7',   os: windows-latest,   payload: serial}
           - {version: nightly,   os: ubuntu-latest,  payload: serial}
     steps:
       - uses: actions/checkout@v2
