### 2022-10-12 255c282
docs&docstrings modified
