### 2022-10-12 SotaYoshida/NuclearToolkit.jl#38 @SotaYoshida
> To be fair, that might be a common misunderstanding :)

You're absolutely right.

>A suggestion: Have you tested this on Windows? Perhaps the same idea can be conveyed by saying "works on a variety of environments, including Mac, linux, and Windows"? This shouldn't result in fewer understanding about hardware.

Not actually. I added the test through GitHub Actions on Windows, and the code works fine.

I will make a pull request and the updated document will be available soon.
I would appreciate it if you could take a look through it.
