+++ src/chiEFTint/matter.jl
@@ -1,14 +1,15 @@
 """
-Reference: K.Hebeler & A.Schwenk, Phys. Rev. C 82, 014314 (2010).
+Reference: 
 """
-function calc_nuclearmatter_in_momspace(chiEFTobj::ChiralEFTobject,to,io)      
+function calc_nuclearmatter_in_momspace(chiEFTobj::ChiralEFTobject,to,io)
+    @assert chiEFTobj.params.n_mesh >= 100 "n_mesh >=100 is recomended for E/A calc."
     n_below_kF = get_n_below_kF(chiEFTobj)
     kF = chiEFTobj.params.kF
     rho = (2 *kF^3 )/ (3*pi^2)
     println(io,"kF (fm^-1) $kF rho (fm^-3) $rho n_below_kF $n_below_kF")
     TperA = 3*hc^2 *kF^2 /(5*2* (Mp+Mn)/2)
     EHF  = calc_EperA_HF(chiEFTobj,n_below_kF)
-    EPT2 = calc_EperA_PT2(chiEFTobj,n_below_kF,to)
+    EPT2 = calc_EperA_PT2(chiEFTobj,to)
     E = TperA + EHF + EPT2
     println(io,"kinetic ",@sprintf("%10.3f",TperA)," EHF/A ",@sprintf("%10.3f",EHF)," EPT2/A ",@sprintf("%15.8f",EPT2),
     "   E(1) ", @sprintf("%10.3f",E-EPT2)," E(1+2) ",@sprintf("%10.3f",E))
@@ -29,25 +30,71 @@ function calc_EperA_HF(chiEFTobj,n_below_kF;verbose=false)
         if l1 != l2; continue;end
         Jfac = 2*J + 1
         LSTfac = 1 - (-1)^(l1+S+1)
-        if LSTfac == 0;continue;end
         tsum = 0.0
+        afac = ifelse(Tz==0,sqrt(2.0),LSTfac/sqrt(2.0))
         for i = 1:n_below_kF
-            k = xr_fm[i]; k2 = k^2 
+	    k = xr_fm[i]; k2 = k^2
             w = wr[i]
             v2 = tV[i,i]; v2n3n = tV_2n3n[i,i]
-            tsum  += k2 * w * Jfac * LSTfac *  (v2 + v2n3n/3)
+            tsum  += k2 * (1-1.5*(k/kF)+ 0.5 * (k/kF)^3 ) * w * Jfac * (v2* afac^2 + v2n3n/3)
         end
         sumarrs[threadid()] += tsum
-        if verbose && abs(tsum) >  1.0
-            println("to HF: L $l1 S $S J $J tsum ",@sprintf("%10.4e",tsum))
+        if verbose && abs(tsum) >  1.0                                                                        
+           println("to HF: L $l1 S $S J $J tsum ",@sprintf("%10.4e",tsum))
         end
     end
-    EHF = sum(sumarrs) * 8 / pi^3 
-    return EHF
+    EHF = sum(sumarrs)
+end
+function calc_EperA_PT2(chiEFTobj::ChiralEFTobject,to)
+    kF = chiEFTobj.params.kF
+    n_mesh = chiEFTobj.params.n_mesh
+    V12mom = chiEFTobj.V12mom
+    V12mom_2n3n = chiEFTobj.V12mom_2n3n
+    pw_channels = chiEFTobj.pw_channels
+    xr_fm = chiEFTobj.xr_fm
+    wr = chiEFTobj.wr
+    nKmesh = 50
+    Ks,wKs = Gauss_Legendre(0,2*kF,nKmesh)
+    EPT2 = 0.0
[…]
