### 2022-10-20 088377b
improved comm221 performance
