### 2022-11-10 5a87563
[Doc:] guide to sample results is added
