+++ .github/workflows/CI.yml
@@ -60,6 +60,16 @@ jobs:
         with:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
+      
+      - name: Cleanup
+        working-directory: ./doc
+        run: rm -rf log_sample_script.txt
+      - name: Run sample_script.jl
+        run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
+      - name: commit changed files
+        uses: stefanzweifel/git-auto-commit-action@v4.2.0
+        with:
+          commit_message: Update log_sample_script.txt
   
   docs:
     name: Documentation
@@ -80,24 +90,6 @@ jobs:
             using NuclearToolkit
             DocMeta.setdocmeta!(NuclearToolkit, :DocTestSetup, :(using NuclearToolkit); recursive=true)
             doctest(NuclearToolkit)'
-      
-      - name: Cleanup
-        working-directory: ./doc
-        run: rm -rf log_sample_script.txt
-      - name: Run sample_script.jl
-        run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
-      - name: commit changed files
-        uses: stefanzweifel/git-auto-commit-action@v4.2.0
-        with:
-          commit_message: Update log_sample_script.txt
 
-    steps:
-      - uses: actions/checkout@v2
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: ${{ matrix.version }}
-          arch: x64
-      - uses: julia-actions/cache@v1
-      - uses: julia-actions/julia-buildpkg@v1
-      - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
+
 
