+++ .github/workflows/CI.yml
@@ -68,7 +68,8 @@ jobs:
         include:
           - {version: '1.7',   os: ubuntu-latest,  payload: mpi}
     name: deploy results of Sample script
-      working-directory: ./doc
+    steps:
+      working-directory: ./example
       run: rm -rf log_sample_script.txt
       - name: Run sample_script.jl
         run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
