+++ .github/workflows/CI.yml
@@ -60,17 +60,9 @@ jobs:
         with:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - {version: '1.7',   os: ubuntu-latest,  payload: mpi}
-    name: deploy results of Sample script
-    steps:
-      working-directory: ./example
-      run: rm -rf log_sample_script.txt
+
+      - working-directory: ./example
+      - run: rm -rf log_sample_script.txt
       - name: Run sample_script.jl
         run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
       - name: commit changed files
