+++ .github/workflows/CI.yml
@@ -61,8 +61,7 @@ jobs:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
 
-      - working-directory: ./example
-      - run: rm -rf log_sample_script.txt
+      - run: rm -rf example/log_sample_script.txt
       - name: Run sample_script.jl
         run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
       - name: commit changed files
