+++ example/optional_parameters.jl
@@ -1,0 +1,14 @@
+###--- ChiralEFT ---
+n_mesh = 50
+emax = 2
+calc_NN = true
+calc_3N = false
+hw = 20.0
+srg = false
+tbme_fmt = "snt.bin"
+chi_order = 3; pottype="em500n3lo"
+
+### --- IMSRG ---
+smax = 500.0
+dsmax = 0.5
+denominatorDelta=0.0
