+++ .github/workflows/draft-pdf.yml
@@ -1,7 +1,7 @@
 on:
   push:
     paths:
-      - "docs/**"
+      - "paper/*"
 
 jobs:
   paper:
