+++ .github/workflows/CI.yml
@@ -60,7 +60,8 @@ jobs:
         with:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
-
+    name: Run sample_script.jl
+    steps:
       - run: rm -rf example/log_sample_script.txt
       - name: Run sample_script.jl
         run: julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
