+++ .github/workflows/CI.yml
@@ -60,6 +60,13 @@ jobs:
         with:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
+  build2:
+  runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - {version: '1.7',   os: ubuntu-latest,  payload: mpi}
     name: Run sample_script.jl
     steps:
       - run: rm -rf example/log_sample_script.txt
