+++ .github/workflows/CI.yml
@@ -78,7 +78,7 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1 
       - name: Run sample_script.jl
         run: git_hash=$(git rev-parse --short "$GITHUB_SHA")
-        run: echo $git_hash > ./example/log_sample_script.txt
+        run: echo $git_hash  > ./example/log_sample_script.txt
         run: julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
       - name: commit changed files
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
