+++ .github/workflows/CI.yml
@@ -77,7 +77,7 @@ jobs:
       - uses: julia-actions/cache@v1
       - uses: julia-actions/julia-buildpkg@v1 
       - name: Run sample_script.jl
-      - run: | 
+        - run: | 
           git_hash=$(git rev-parse --short "$GITHUB_SHA")
           run: echo $git_hash  > ./example/log_sample_script.txt
           run: julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
