+++ .github/workflows/CI.yml
@@ -77,9 +77,9 @@ jobs:
       - uses: julia-actions/cache@v1
       - uses: julia-actions/julia-buildpkg@v1 
       - run: | 
-          git_hash=$(git rev-parse --short "$GITHUB_SHA")
-          echo $git_hash  > ./example/log_sample_script.txt
-          julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
+        git_hash=$(git rev-parse --short "$GITHUB_SHA")
+        echo $git_hash  > ./example/log_sample_script.txt
+        julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
       - name: commit changed files
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
         with:
