+++ .github/workflows/CI.yml
@@ -79,7 +79,7 @@ jobs:
       - run: | 
           git_hash=$(git rev-parse --short "$GITHUB_SHA")
           echo git_hash $git_hash
-          echo $git_hash  > ./example/log_sample_script.txt
+          echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
           julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
       - name: commit changed files
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
