+++ docs/src/index.md
@@ -21,7 +21,7 @@ You can try the package by 1. or 2.:
       - calculating NN potential from Chiral EFT
       - HFMBPT(3) and IMSRG/VS-IMSRG(2) calculation with it
       - shell-model calculations with the effective interaction derived by VS-IMSRG
-    The corresponding results with the latest release can be found in [here](https://github.com/SotaYoshida/NuclearToolkit.jl/blob/main/sample_results.dat)
+    An expected results using the latest dev branch can be found [here](https://github.com/SotaYoshida/NuclearToolkit.jl/blob/develop/example/log_sample_script.txt).
 2. Try sample codes in [HowToUse](howtouse) page.
 
 
