### 2022-11-10 591de93
Update log_sample_script.txt
