### 2022-11-10 051e6ee
Update log_sample_script.txt
