+++ .github/workflows/CI.yml
@@ -61,6 +61,7 @@ jobs:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
   build2:
+    if: ${{ branches == 'develop' }}
     runs-on: ubuntu-latest
     strategy:
       fail-fast: false
@@ -81,7 +82,6 @@ jobs:
           echo git_hash $git_hash
           echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
           julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
-      - if: ${{ branches == 'develop' }}
       - name: commit changed files
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
         with:
