+++ .github/workflows/CI_pullrequest.yml
@@ -1,4 +1,4 @@
-name: CI
+name: CI_PR
 on:
   pull_request:
     branches:
