+++ .github/workflows/CI.yml
@@ -68,6 +68,7 @@ jobs:
           - {version: '1.7',   os: ubuntu-latest}
     name: Run sample_script.jl
     steps:     
+      - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
