### 2022-11-10 57809cf
Update log_sample_script.txt
