### 2022-11-10 4f930eb
docs are updated for new CI with example/sample_script.jl
