+++ optional_parameters.jl
@@ -1,7 +1,7 @@
 ###--- ChiralEFT ---
 n_mesh = 50
 emax = 4
-chi_order = 4 #0:LO 1:NLO 2:NNLO 3:N3LO
+chi_order = 3
 calc_NN = true
 calc_3N = false 
 hw = 20.0
