### 2022-11-10 4025ad9
docs fixed for new CI using example/sample_script.jl
