+++ docs/src/index.md
@@ -13,16 +13,17 @@ julia>]add NuclearToolkit
 
 You can try the package by 1. or 2.:
 
-1. clone the repository and run `example/sample_script.jl` in the repository like  
-    ```
-      $ julia -t 8 example/sample_script.jl
-    ```
+1. clone the repository and run `example/sample_script.jl` in the repository like
+
+    `$ julia -t 8 example/sample_script.jl`
+
     This performs:
-      - calculating NN potential from Chiral EFT
+      - calculation of NN potential from Chiral EFT
       - HFMBPT(3) and IMSRG/VS-IMSRG(2) calculation with it
       - shell-model calculations with the effective interaction derived by VS-IMSRG
     An expected results using the latest dev branch can be found [here](https://github.com/SotaYoshida/NuclearToolkit.jl/blob/develop/example/log_sample_script.txt).
 2. Try sample codes in [HowToUse](howtouse) page.
+
 
 
 Please make sure to use the latest version of the package. Update can be done with 
