+++ .github/workflows/CI_pullrequest.yml
@@ -81,6 +81,7 @@ jobs:
       - run: | 
           git_hash=$(git rev-parse --short "$GITHUB_SHA")
           echo git_hash $git_hash
+          ls *
           echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
           julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
           git add examples/log_sample_script.txt
