+++ .github/workflows/CI.yml
@@ -2,6 +2,7 @@ name: CI
 on:
   push:
     branches:
+      - main
       - develop
     tags: '*'
     exclude: "log_sample_script.txt"
