+++ .github/workflows/CI_pullrequest.yml
@@ -84,7 +84,7 @@ jobs:
           echo running on branch ${GITHUB_REF##*/}    
           echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
           julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
-          git add examples/log_sample_script.txt
+          git add example/log_sample_script.txt
       - name: commit changed logfile
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
         with:
