### 2022-11-10 384101e
Update log_sample_script.txt
