+++ .github/workflows/CI_pullrequest.yml
@@ -59,32 +59,32 @@ jobs:
           mpi: ${{ matrix.mpi }}
       - run: mpiexec -n 4 julia --project="@." ./test/mpirun_test.jl
 
-  build2:    
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-          - {version: '1.7',   os: ubuntu-latest}
-    name: Run sample_script.jl
-    steps:     
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 0 
-      - uses: julia-actions/setup-julia@v1
-        with:
-          version: ${{ matrix.version }}
-          arch: x64
-      - uses: julia-actions/cache@v1
-      - uses: julia-actions/julia-buildpkg@v1 
-      - run: | 
-          git_hash=$(git rev-parse --short "$GITHUB_SHA")
-          echo git_hash $git_hash     
-          echo running on branch ${GITHUB_REF##*/}    
-          echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
-          julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
-          git add example/log_sample_script.txt
-          
+  # build2:    
+  #   runs-on: ubuntu-latest
+  #   strategy:
+  #     fail-fast: false
+  #     matrix:
+  #       include:
+  #         - {version: '1.7',   os: ubuntu-latest}
+  #   name: Run sample_script.jl
+  #   steps:     
+  #     - uses: actions/checkout@v2
+  #       with:
+  #         fetch-depth: 0 
+  #     - uses: julia-actions/setup-julia@v1
+  #       with:
+  #         version: ${{ matrix.version }}
+  #         arch: x64
+  #     - uses: julia-actions/cache@v1
+  #     - uses: julia-actions/julia-buildpkg@v1 
+  #     - run: | 
+  #         git_hash=$(git rev-parse --short "$GITHUB_SHA")
+  #         echo git_hash $git_hash     
+  #         echo running on branch ${GITHUB_REF##*/}    
+  #         echo git_hash:[$git_hash]  > ./example/log_sample_script.txt
+  #         julia --project="@." -t 4 ./example/sample_script.jl >> ./example/log_sample_script.txt
+  #         git add example/log_sample_script.txt
+
       # - name: commit changed logfile
       #   uses: stefanzweifel/git-auto-commit-action@v4.2.0
       #   with:
