### 2022-11-10 60076d6
Update log_sample_script.txt
