+++ example/log_sample_script.txt
@@ -1,1 +1,288 @@
+git_hash:[141e1b2]
+@emax 4 hitCG 29351 dWS <  11.31 MB   9j(1415) <   2.51 MB   HOB (5179) <   0.31 MB
+target: O16 Ref. => Z=8 N=8 E:    -132.9506  = E1b    363.70  + E2b   -496.65   (    -55.16  -367.32   -74.18)
+EMP2    -23.70628 1b    -0.000 pp    -1.437 pn   -20.815 nn    -1.454
+EMP3      0.37645 pp    -0.111 hh    -1.895 ph     2.383
+E_HF   -132.95058  E_MBPT(3) =    -156.2804  Eexp:     -127.619
 
+parameters in optional_parameters.jl will be used.
+ step:        s             E0   ||Omega_1||   ||Omega_2||     ||Eta_1||     ||Eta_2||    Ncomm.  nwritten
+     0    0.000  -132.95057781  0.000000e+00  0.000000e+00  7.576852e-17  1.410292e+00         0         0
+     1    0.500  -150.64837608  3.788426e-17  7.051459e-01  4.457285e-02  6.465559e-01         8         0
+     2    1.000  -154.78704837  2.228643e-02  3.232779e-01  5.387093e-02  3.441643e-01        15         1
+     3    1.500  -156.02709506  2.693546e-02  1.720822e-01  5.200791e-02  1.968159e-01        21         2
+     4    2.000  -156.46234613  5.291574e-02  2.697514e-01  4.609815e-02  1.192463e-01        28         2
+     5    2.500  -156.63786069  2.304908e-02  5.962317e-02  3.915110e-02  7.603387e-02        33         3
+     6    3.000  -156.71754530  4.261375e-02  9.740456e-02  3.246552e-02  5.082054e-02        38         3
+     7    3.500  -156.75734646  5.883470e-02  1.223106e-01  2.655233e-02  3.554179e-02        44         3
+     8    4.000  -156.77877266  7.209319e-02  1.393224e-01  2.156017e-02  2.604640e-02        51         3
+     9    4.500  -156.79104487  8.284970e-02  1.513341e-01  1.745834e-02  2.007233e-02        58         3
+    10    5.000  -156.79848130  9.154892e-02  1.600942e-01  1.414576e-02  1.629448e-02        65         3
+    11    5.500  -156.80324376  9.858457e-02  1.666945e-01  1.150104e-02  1.388250e-02        72         3
+    12    6.000  -156.80646861  1.042894e-01  1.718341e-01  9.405852e-03  1.230019e-02        79         3
+    13    6.500  -156.80877531  1.089366e-01  1.759706e-01  7.754598e-03  1.120726e-02        86         3
+    14    7.000  -156.81051148  1.127468e-01  1.794075e-01  6.457349e-03  1.039706e-02        93         3
+    15    7.500  -156.81187738  1.158958e-01  1.823489e-01  5.439592e-03  9.749739e-03       100         3
+    16    8.000  -156.81299138  1.185223e-01  1.849332e-01  4.640660e-03  9.198211e-03       107         3
+    17    8.500  -156.81392516  1.207286e-01  1.872546e-01  4.015938e-03  8.705218e-03       113         3
+    18    9.000  -156.81472446  1.226094e-01  1.893775e-01  3.520493e-03  8.251996e-03       119         3
+    19    9.500  -156.81541870  1.242303e-01  1.913467e-01  3.124629e-03  7.827798e-03       125         3
+    20   10.000  -156.81602785  1.256424e-01  1.931932e-01  2.805005e-03  7.426899e-03       131         3
+    21   10.500  -156.81656616  1.268858e-01  1.949382e-01  2.543557e-03  7.046196e-03       137         3
[…]
