### 2022-11-10 beabcef
Update log_sample_script.txt
