+++ example/log_sample_script.txt
@@ -1,6 +1,5 @@
-git_hash:[51c4609]
-@emax 4 hitCG 29351 dWS <  11.31 MB   9j(1419) <   2.51 MB   HOB (5179) <   0.31 MB
-
+git_hash:[d7f8d99]
+@emax 4 hitCG 29351 dWS <  11.31 MB   9j(1427) <   2.51 MB   HOB (5179) <   0.31 MB
 target: O16 Ref. => Z=8 N=8 E:    -132.9506  = E1b    363.70  + E2b   -496.65   (    -55.16  -367.32   -74.18)
 EMP2    -23.70628 1b    -0.000 pp    -1.437 pn   -20.815 nn    -1.454
 EMP3      0.37645 pp    -0.111 hh    -1.895 ph     2.383
@@ -8,8 +7,8 @@ E_HF   -132.95058  E_MBPT(3) =    -156.2804  Eexp:     -127.619
 
 parameters in optional_parameters.jl will be used.
  step:        s             E0   ||Omega_1||   ||Omega_2||     ||Eta_1||     ||Eta_2||    Ncomm.  nwritten
-     0    0.000  -132.95057781  0.000000e+00  0.000000e+00  7.576852e-17  1.410292e+00         0         0
-     1    0.500  -150.64837608  3.788426e-17  7.051459e-01  4.457285e-02  6.465559e-01         8         0
+     0    0.000  -132.95057781  0.000000e+00  0.000000e+00  1.001482e-16  1.410292e+00         0         0
+     1    0.500  -150.64837608  5.007412e-17  7.051459e-01  4.457285e-02  6.465559e-01         8         0
      2    1.000  -154.78704837  2.228643e-02  3.232779e-01  5.387093e-02  3.441643e-01        15         1
      3    1.500  -156.02709506  2.693546e-02  1.720822e-01  5.200791e-02  1.968159e-01        21         2
      4    2.000  -156.46234613  5.291574e-02  2.697514e-01  4.609815e-02  1.192463e-01        28         2
