### 2022-11-10 6f170b1
Update log_sample_script.txt
