+++ .github/workflows/CI.yml
@@ -75,9 +75,8 @@ jobs:
       - uses: julia-actions/cache@v1
       - uses: julia-actions/julia-buildpkg@v1 
       - run: | 
-          git_hash=$(git rev-parse --short "$GITHUB_SHA")
-          echo git_hash $git_hash          
-          julia --project="@." ./example/sample_script.jl >> ./example/log_sample_script.txt
+          git_hash=$(git rev-parse --short "$GITHUB_SHA")          
+          julia --project="@." -t 4 ./example/sample_script.jl > ./example/log_sample_script.txt
       - name: commit changed logfile
         uses: stefanzweifel/git-auto-commit-action@v4.2.0
         with:
