+++ src/chiEFTint/main_chiEFTint.jl
@@ -43,6 +43,7 @@ function make_chiEFTint(;is_show=false,itnum=1,writesnt=true,nucs=[],optimizer="
     end
     if io != stdout; close(io);end
     show_TimerOutput_results(to;tf=is_show)
+    return true
 end
 
 """
