### 2022-11-10 9e1ac3c
update log_sample
