+++ example/log_sample_script.txt
@@ -1,5 +1,4 @@
 @emax 4 hitCG 29351 dWS <  11.31 MB   9j(1430) <   2.51 MB   HOB (5179) <   0.31 MB
-
 target: O16 Ref. => Z=8 N=8 E:    -132.9506  = E1b    363.70  + E2b   -496.65   (    -55.16  -367.32   -74.18)
 EMP2    -23.70628 1b    -0.000 pp    -1.437 pn   -20.815 nn    -1.454
 EMP3      0.37645 pp    -0.111 hh    -1.895 ph     2.383
