### 2022-11-11 #4694 @villaa
I now recommend this paper for publication, after the authors work to improve the documentation and example running.
