### 2022-11-11 #4694 @rkurchin
Some minor editorial comments:
* line 31: what is meant by "FORTRAN/Fortran" here? I believe [standard practice](https://fortranwiki.org/fortran/show/FAQ) is now to spell it as "Fortran" so I would suggest changing to that
* 34-35: rephrase to "This is achieved thanks to the high readability and portability of the Julia programming language" and add space before citation
* 36: add period at the end of the sentence.
