+++ paper/paper.bib
@@ -25,7 +25,7 @@ number = "2",
 pages = "517 - 599",
 year = "2001",
 issn = "0146-6410",
-doi = "https://doi.org/10.1016/S0146-6410(01)00159-4",
+doi = "10.1016/S0146-6410(01)00159-4",
 url = "http://www.sciencedirect.com/science/article/pii/S0146641001001594",
 author = {B A Brown}
 }
@@ -206,7 +206,7 @@ volume = "120",
 pages = "115 - 118",
 year = "2014",
 issn = "0090-3752",
-doi = "https://doi.org/10.1016/j.nds.2014.07.022"
+doi = "10.1016/j.nds.2014.07.022"
 }
 @misc{KSHELL1,
 Author = {Noritaka Shimizu},
@@ -223,7 +223,7 @@ volume = "244",
 pages = "372 - 384",
 year = "2019",
 issn = "0010-4655",
-doi = "https://doi.org/10.1016/j.cpc.2019.06.011",
+doi = "10.1016/j.cpc.2019.06.011",
 url = "http://www.sciencedirect.com/science/article/pii/S0010465519301985",
 author = "Noritaka Shimizu and Takahiro Mizusaki and Yutaka Utsuno and Yusuke Tsunoda"
 }
