+++ example/log_sample_script.txt
@@ -6,8 +6,8 @@ E_HF   -132.95058  E_MBPT(3) =    -156.2804  Eexp:     -127.619
 
 parameters in optional_parameters.jl will be used.
  step:        s             E0   ||Omega_1||   ||Omega_2||     ||Eta_1||     ||Eta_2||    Ncomm.  nwritten
-     0    0.000  -132.95057781  0.000000e+00  0.000000e+00  1.001482e-16  1.410292e+00         0         0
-     1    0.500  -150.64837608  5.007412e-17  7.051459e-01  4.457285e-02  6.465559e-01         8         0
+     0    0.000  -132.95057781  0.000000e+00  0.000000e+00  7.576852e-17  1.410292e+00         0         0
+     1    0.500  -150.64837608  3.788426e-17  7.051459e-01  4.457285e-02  6.465559e-01         8         0
      2    1.000  -154.78704837  2.228643e-02  3.232779e-01  5.387093e-02  3.441643e-01        15         1
      3    1.500  -156.02709506  2.693546e-02  1.720822e-01  5.200791e-02  1.968159e-01        21         2
      4    2.000  -156.46234613  5.291574e-02  2.697514e-01  4.609815e-02  1.192463e-01        28         2
