### 2022-11-12 d939ae8
betadecay.jl and amedata.jl modified
