+++ src/ShellModel/betadecay.jl
@@ -1,12 +1,26 @@
 const K = 6144.0
 const lam_gav = 1.2701
 
-function eval_betadecay_from_kshell_log(fns_sum_parent,fns_sum_daughter,fns_GT,fns_FF,parentJpi;pnuc="",dnuc="")
+"""
+    eval_betadecay_from_kshell_log(fns_sum_parent::Vector{String},fns_sum_daughter::Vector{String},fns_GT::Vector{String},fns_FF::Vector{String},parentJpi::String;pnuc="",dnuc="",verbose=false)
+
+# Arguments
+- `fns_sum_parent::Vector{String}` vector of paths to KSHELL-summary files for the parent nucleus
+- `fns_sum_daughter::Vector{String}` vector of paths to KSHELL-summary files for the daughter nucleus
+- `fns_GT::Vector{String}` vector of paths to KSHELL-log files for Gamow-Teller transitions.
+- `fns_FF::Vector{String}` vector of paths to KSHELL-log files for First-Forbidden transitions.
+
+# Optional Arguments
+- `pnuc::String` to specify the parent nuclei explicitly
+- `dnuc::String` to specify the daughter nuclei explicitly
+- `verbose::Bool` option to show GT/FF transitions for each state
+"""
+function eval_betadecay_from_kshell_log(fns_sum_parent::Vector{String},fns_sum_daughter::Vector{String},fns_GT::Vector{String},fns_FF::Vector{String},parentJpi::String;pnuc="",dnuc="",verbose=false)
     qfactors = get_quenching_factors()
     parent = read_kshell_summary(fns_sum_parent;nuc=pnuc,targetJpi=parentJpi)
     daughter = read_kshell_summary(fns_sum_daughter;nuc=dnuc)
     qvals = get_qval(parent,daughter,parentJpi)
-    eval_logft_f(parent,daughter,qvals,fns_GT,fns_FF,qfactors)
+    eval_logft_f(parent,daughter,qvals,fns_GT,fns_FF,qfactors;verbose=verbose)
 end
 
 mutable struct quenching_factors
@@ -16,19 +30,6 @@ mutable struct quenching_factors
     qx::Float64
     qu::Float64
     qz::Float64
-end
-
-function get_quenching_factors(qtype="")
-    if qtype == "" || qtype == "SY" 
-        # S. Yoshida et al., PRC 97, 054321(2018).
-        return quenching_factors(0.74,1.0,1.7,1.0,1.0,0.55) 
-    elseif qtype == "W" || qtype == "Warburton" 
-        # E. Warburton, J. Becker, B. Brown, and D. Millener, Ann. Phys. 187, 471 (1988).
-        return quenching_factors(1.0,1.1,1.5,1.0,1.0,0.51)
-    else
-        @warn "qtype=$qtype is not supported! SY value will be used"
-        return quenching_factors(0.74,1.0,1.7,1.0,1.0,0.55) 
-    end
 end
 
 """
@@ -52,29 +53,44 @@ function Ecoulomb_SM(Z,N)
     return 0.7 * (Z*(Z-1) -0.76 * (Z*(Z-1))^(2/3)) / Rc
 end
 
-function get_qval(parent,daughter,parentJpi)
+function get_quenching_factors(qtype="")
+    if qtype == "" || qtype == "SY" 
+        # S. Yoshida et al., PRC 97, 054321(2018).
[…]
