+++ docs/src/ShellModel.md
@@ -6,6 +6,8 @@ Functions for shell-model calculations
 - [transit.jl](https://github.com/SotaYoshida/NuclearToolkit.jl/tree/main/src/ShellModel/transit.jl): EM transitions
 - [input_int_snt.jl](https://github.com/SotaYoshida/NuclearToolkit.jl/tree/main/src/ShellModel/input_int_snt.jl): I/O stuffs
 - [eigenvector_continuation.jl](https://github.com/SotaYoshida/NuclearToolkit.jl/tree/main/src/ShellModel/eigenvector_continuation.jl): eigenvector continuation to sample and construct shell-model wavefunctions
+- [KSHELL.jl](https://github.com/SotaYoshida/NuclearToolkit.jl/tree/main/src/ShellModel/KSHELL.jl): dealing with KSHELL outputs
+- [betadecay.jl](https://github.com/SotaYoshida/NuclearToolkit.jl/tree/main/src/ShellModel/betadecay.jl): evaluate beta-decay properties using KSHELL log/summary
 
 ```@autodocs
 Modules = [NuclearToolkit]
