### 2022-11-12 #4694 @SotaYoshida
@rkurchin
I have fixed DOIs and papers, and then released the new version, [v.0.2.5](https://github.com/SotaYoshida/NuclearToolkit.jl/releases/tag/v0.2.5)
The DOI of the Zenodo archive is [10.5281/zenodo.7314918](https://doi.org/10.5281/zenodo.7314918).

Please let me know if i missed anything. Thanks.
