### 2022-11-14 #4694 @rkurchin
Two more tiny fixes on paper:
* line 26: should begin "In the nuclear physics community..."
* you're still missing a space before the citation on line 35

After that we should be good to go!
