### 2022-11-14 #4694 @SotaYoshida
@rkurchin Thanks!
I have fixed the two points. (I made a josspaper branch and then merged to the main branch)
