+++ paper/paper.bib
@@ -116,7 +116,7 @@ keywords = "Nucleon\UTF{2013}nucleon scattering "
   url = {http://link.aps.org/doi/10.1103/RevModPhys.81.1773}
 }
 @ARTICLE{LENPICrev2,
-AUTHOR={Epelbaum, Evgeny and Krebs, Hermann and Reinert, Patrick},
+AUTHOR={Epelbaum, E. and Krebs, H. and Reinert, P.},
 TITLE={High-Precision Nuclear Forces From Chiral EFT: State-of-the-Art, Challenges, and Outlook},
 JOURNAL={Frontiers in Physics},
 VOLUME={8}, 
@@ -191,7 +191,7 @@ doi = {10.1016/j.cpc.2013.07.022},
 url = {https://www.sciencedirect.com/science/article/pii/S0010465513002518},
 }
 @misc{BIGSTICK2,
-Title = {BIGSTICK: A flexible configuration-interaction shell-model code},
+Title = {{BIGSTICK}: A flexible configuration-interaction shell-model code},
 year = {2018},
 journal = {arXiv preprint arXiv:1801.08432},
 author = {Calvin W. Johnson and W. Erich Ormand and Kenneth S. McElvain and Hongzhang Shan},
@@ -210,7 +210,7 @@ doi = "10.1016/j.nds.2014.07.022"
 }
 @misc{KSHELL1,
 Author = {Noritaka Shimizu},
-Title = {Nuclear shell-model code for massive parallel computation, "KSHELL"},
+Title = {Nuclear shell-model code for massive parallel computation, ``KSHELL''},
 Year = {2013},
 journal = {arXiv preprint arXiv:1310.5431},
 doi = {10.48550/arXiv.1310.5431},
