### 2022-11-18 #4694 @kyleniemeyer
Hi @SotaYoshida, I made a small fix to the `.bib` file, could you merge that? https://github.com/SotaYoshida/NuclearToolkit.jl/pull/63
